### 2021-02-19 #430 @wlandau
stantargets: reproducible Stan pipelines at scale
Submitting Author: Will Landau (@wlandau)
Repository:  https://github.com/wlandau/stantargets
Version submitted:   0.0.0.9000
Version accepted: TBD

---

```
Package: stantargets
Title: Targets for Stan Workflows
Description: Bayesian data analysis usually incurs long runtimes
  and cumbersome custom code. A specialized pipeline toolkit for
  Bayesians, the 'stantargets' R package leverages
  'targets' and 'cmdstanr' to ease these burdens.
  'stantargets' makes it super easy to set up useful scalable
  Stan pipelines that automatically parallelize the computation
  and skip expensive steps when the results are already up to date.
  Minimal custom code is required, and there is no need to manually
  configure branching, so usage is much easier than 'targets' alone.
  'stantargets' can access all of 'cmdstanr''s major algorithms
  (MCMC, variational Bayes, and optimization) and it supports
  both single-fit workflows and multi-rep simulation studies.
  For the statistical methodology, please refer to 'Stan' documentation
  (Stan Development Team 2020) <https://mc-stan.org/>.
Version: 0.0.0.9000
  person(
    given = c("William", "Michael"),
    family = "Landau",
    email = "will.landau@gmail.com",
    comment = c(ORCID = "0000-0003-1878-3253")
  ),
  person(
    family = "Eli Lilly and Company",
    role = "cph"
  ))
  cmdstanr (>= 0.2.0),
  digest (>= 0.6.21),
  fst (>= 0.9.4),
  posterior (>= 0.1.2),
  purrr (>= 0.3.4),
  qs (>= 0.14.1),
  rlang (>= 0.4.8),
  targets (>= 0.0.1),
  tarchetypes (>= 0.0.1),
  tibble (>= 3.0.4),
  tools
  fs (>= 1.5.0),
  knitr (>= 1.28),
  R.utils (>= 2.10.1),
  rmarkdown (>= 2.1),
  visNetwork (>= 2.0.9),
  withr (>= 2.1.2)
  stan-dev/cmdstanr,
  stan-dev/posterior
SystemRequirements: CmdStan >= 2.25.0
```

`stantargets` is very similar to `jagstargets` (#425). `stantargets` leverages the existing workflow automation capabilities of [`targets`](https://docs.ropensci.org/targets) to orchestrate computation and skip up-to-date tasks in Bayesian data analysis pipelines. `stantargets` reduces the burden of user-side custom code that [`targets`](https://docs.ropensci.org/targets) would otherwise require, which helps free statisticians to focus more on the models and less on the software engineering.

`stantargets` is for Bayesian statisticians who develop and run Stan models. Example workflows range from individual analyses of clinical data to large-scale simulation-based calibration studies for validation.

[`targets`](https://docs.ropensci.org/targets) already provides the same kind of workflow automation, but it requires more custom code to set up a workflow. `stantargets` uses specialized domain knowledge to make this process easier. Packages `rstan` and `cmdstanr` interface with Stan but do not provide the same kind of workflow automation. In light of the [recent preprint by Gelman et al. (2020)](https://arxiv.org/abs/2011.01808), I believe the Stan Development Team would be very interested in this kind of workflow automation.

N/A

N/A

__If JOSS is still an option, I would like to publish there. I have prepared a manuscript at https://github.com/wlandau/stantargets/blob/main/inst/paper.md.__
