### 2021-02-22 #430 @melvidoni
### Editor checks:

- [X] **Automated tests:** Package has a testing suite and is tested via a CI service.

---

#### Editor comments

Thank you for submitting it! I'll be your handling editor this time. `Goodpractice` is very happy about it:

```
♥ Aha! Fantastic package! Keep up the prime work!
```

There are no spelling checks, and `covr` is also happy. Good job!

I will be searching for reviewers then.

---

Reviewers: TBD
Due date: TBD
