### 2021-03-03 #430 @wlandau
Thanks for the update, @melvidoni. I do understand this may take some time.

> And yes, I have your master list of suggestions.

Do you mean the folks I mentioned at https://github.com/ropensci/software-review/issues/401#issuecomment-711036834 for `targets`? I actually think about `stantargets` a bit differently here because of its focus on Bayesian Statistics. For my own learning, it would be amazing to include someone from the [Stan Development Team](https://mc-stan.org/about/team/). If not, then I think practicing Bayesian statisticians, particularly Stan users, would have a lot to add.
