### 2021-03-03 #430 @wlandau
> Yes, that list.

Oh good, glad we are on the same page. Sorry, I may have misremembered the earlier parts of this thread.

> However, given the current world circumstances, I am afraid that finding reviewers may take longer than expected.

I completely understand.

For what it's worth, dependency packages `cmdstanr` and `posterior` are not yet on CRAN, so this review process is extremely unlikely to bottleneck any of my plans. And unlike `targets`, I am not developing any packages that depend on `stantargets`.
