### 2021-03-04 #430 @sakrejda
Hi Melina, I am interested and the three-week time frame is reasonable for me.  I'm not the right person to say much about whether the package aligns well with how `targets` itself is commonly used but should be able to provide feedback otherwise.  If you haven't asked @tjmahr yet he might be interested in reviewing.
