+++ R/tar_stan_example_data.R
@@ -8,13 +8,14 @@
 #' @examples
 #' tar_stan_example_data()
 tar_stan_example_data <- function(n = 10L) {
-  true_beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
+  beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
   x <- seq(from = -1, to = 1, length.out = n)
-  y <- stats::rnorm(n, x * true_beta, 1)
+  y <- stats::rnorm(n, x * beta, 1)
   list(
     n = n,
     x = x,
     y = y,
-    true_beta = true_beta
+    true_beta = true_beta,
+    .join_data = list(beta = beta)
   )
 }
