+++ tests/testthat/test-tar_stan_mcmc.R
@@ -3,8 +3,7 @@
 targets::tar_test("tar_stan_mcmc(compile = \"original\")", {
   skip_on_cran()
   skip_if_not_installed("dplyr")
-  tar_stan_example_file(path = "a.stan")
-  tar_stan_example_file(path = "b.stan")
+  restore_compiled_models()
   targets::tar_script({
     tar_option_set(memory = "transient", garbage_collection = TRUE)
     list(
@@ -134,6 +133,7 @@ targets::tar_test("tar_stan_mcmc(compile = \"original\")", {
 targets::tar_test("tar_stan_mcmc(compile = \"copy\") with custom summaries", {
   skip_on_cran()
   skip_if_not_installed("dplyr")
+  skip_compile_copy()
   tar_stan_example_file(path = "a.stan")
   tar_stan_example_file(path = "b.stan")
   targets::tar_script({
