+++ tests/testthat/test-tar_stan_mcmc_rep_summary.R
@@ -15,8 +15,8 @@ targets::tar_test("tar_stan_mcmc_rep_summary(compile = \"original\")", {
         quiet = TRUE,
         refresh = 0,
         init = 1,
-        iter_sampling = 1000,
-        iter_warmup = 500,
+        iter_sampling = 20,
+        iter_warmup = 10,
         chains = 4,
         batches = 2,
         reps = 2,
@@ -106,8 +106,8 @@ targets::tar_test("tar_stan_mcmc_rep_summary(compile = \"original\")", {
         quiet = TRUE,
         refresh = 0,
         init = 1,
-        iter_sampling = 1000,
-        iter_warmup = 500,
+        iter_sampling = 20,
+        iter_warmup = 10,
         chains = 4,
         batches = 2,
         reps = 2,
