+++ tests/testthat/test-tar_stan_mcmc_rep_diagnostics.R
@@ -189,7 +189,7 @@ targets::tar_test("tar_stan_mcmc_rep_diagnostics(compile = \"copy\")", {
   expect_equal(length(out), 2L)
   out <- out[[2]]
   expect_true(is.list(out))
-  expect_equal(length(out), 4L)
+  expect_equal(length(out), 5L)
   expect_equal(out$n, 10L)
   expect_equal(length(out$x), 10L)
   expect_equal(length(out$y), 10L)
