+++ man/tar_stan_mle_rep_summary.Rd
@@ -53,8 +53,7 @@ tar_stan_mle_rep_summary(
 )
 }
 \arguments{
-\item{name}{Symbol, base name for the collection of targets.
-Serves as a prefix for target names.}
+\item{name}{Symbol, name of the target.}
 
 \item{stan_files}{Character vector of paths to known existing Stan model
 files created before running the pipeline.}
@@ -87,11 +86,6 @@ method.}
 compilation be suppressed? The default is \code{TRUE}, but if you encounter an
 error we recommend trying again with \code{quiet=FALSE} to see more of the
 output.}
-
-\item{log}{Character of length 1, file path to write the stdout stream
-of the model when it runs. Set to \code{NULL} to print to the console.
-Set to \code{R.utils::nullfile()} to completely suppress all output.
-Does not apply to messages, warnings, or errors.}
 
 \item{dir}{(string) The path to the directory in which to store the CmdStan
 executable (or \code{.hpp} file if using \verb{$save_hpp_file()}). The default is the
