+++ man/tar_stan_mcmc_rep_run.Rd
@@ -83,11 +83,6 @@ compilation be suppressed? The default is \code{TRUE}, but if you encounter an
 error we recommend trying again with \code{quiet=FALSE} to see more of the
 output.}
 
-\item{log}{Character of length 1, file path to write the stdout stream
-of the model when it runs. Set to \code{NULL} to print to the console.
-Set to \code{R.utils::nullfile()} to completely suppress all output.
-Does not apply to messages, warnings, or errors.}
-
 \item{dir}{(string) The path to the directory in which to store the CmdStan
 executable (or \code{.hpp} file if using \verb{$save_hpp_file()}). The default is the
 same location as the Stan program.}
@@ -287,14 +282,6 @@ check the results against some "true value" in the data. See the
 
 \item{variables}{(character vector) The variables to read in. If \code{NULL} (the
 default) then all variables are included.}
-
-\item{summaries}{Optional list of summary functions passed to \code{...} in
-\code{posterior::summarize_draws()} through \verb{$summary()}
-on the \code{CmdStanFit} object.}
-
-\item{summary_args}{Optional list of summary function arguments passed to
-\code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
-on the \code{CmdStanFit} object.}
 }
 \value{
 A data frame of posterior summaries.
