+++ man/tar_stan_summary.Rd
@@ -1,17 +1,16 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/tar_stan_summary.R, R/utils_rep.R
+% Please edit documentation in R/tar_stan_summary.R
 \name{tar_stan_summary}
 \alias{tar_stan_summary}
 \title{One summary of a \code{CmdStanFit} object.}
 \usage{
 tar_stan_summary(fit, data, summaries, summary_args, variables)
-
-tar_stan_summary(fit, data, summaries, summary_args, variables)
 }
 \arguments{
-\item{fit}{A Stan fit object.}
+\item{fit}{Symbol, name of a \code{CmdStanFit} object or an upstream target
+that returns a \code{CmdStanFit} object.}
 
-\item{data}{List, Stan dataset.}
+\item{data}{Code to generate the \code{data} for the Stan model.}
 
 \item{summaries}{Optional list of summary functions passed to \code{...} in
 \code{posterior::summarize_draws()} through \verb{$summary()}
@@ -31,13 +30,9 @@ as described at \url{https://books.ropensci.org/targets/}.
 Please see the design specification at
 \url{https://books.ropensci.org/targets-design/}
 to learn about the structure and composition of target objects.
-
-A data frame of user-friendly Stan output.
 }
 \description{
 Target to run the \verb{$summary()} method of a \code{CmdStanFit} object.
-
-Not a user-side function. Do not invoke directly.
 }
 \details{
 \code{\link[=tar_stan_mcmc]{tar_stan_mcmc()}} etc. with \code{summary = TRUE} already gives you a
@@ -70,4 +65,3 @@ targets::tar_manifest()
 })
 }
 }
-\keyword{internal}
