+++ R/tar_stan_gq.R
@@ -133,6 +133,7 @@ tar_stan_gq <- function(
   )
   command_summary <- tar_stan_summary_call(
     sym_fit = sym_gq,
+    sym_data = sym_data,
     summaries = substitute(summaries),
     summary_args = substitute(summary_args),
     variables = variables
