+++ tests/testthat/test-tar_stan_gq.R
@@ -46,6 +46,8 @@ targets::tar_test("tar_stan_gq(compile = \"original\")", {
      "gq_file_a", "gq_gq_a",
      "gq_file_b", "gq_gq_b",
      "gq_gq_a", "gq_draws_a",
+     "gq_data", "gq_summary_a",
+     "gq_data", "gq_summary_b",
      "gq_gq_a", "gq_summary_a",
      "gq_gq_b", "gq_draws_b",
      "gq_gq_b", "gq_summary_b",
@@ -198,6 +200,8 @@ targets::tar_test("tar_stan_gq(compile = \"copy\") with custom summaries", {
     "gq_gq_a", "gq_summary_a",
     "gq_gq_b", "gq_draws_b",
     "gq_gq_b", "gq_summary_b",
+    "gq_data", "gq_summary_a",
+    "gq_data", "gq_summary_b",
     "gq_lines_a", "gq_gq_a",
     "gq_lines_b", "gq_gq_b",
     "model_data", "gq_data",
