### 2021-03-09 3ea07e2
Hide .join_data from Stan
