+++ R/tar_stan_mle_rep.R
@@ -374,8 +374,10 @@ tar_stan_mle_rep_run_rep <- function(
   summaries,
   summary_args
 ) {
+  stan_data <- data
+  stan_data$.join_data <- NULL
   fit <- model$optimize(
-    data = data,
+    data = stan_data,
     seed = seed,
     refresh = refresh,
     init = init,
