+++ R/tar_stan_mle.R
@@ -317,8 +317,10 @@ tar_stan_mle_run <- function(
   if (is.null(seed)) {
     seed <- abs(targets::tar_seed()) + 1L
   }
+  stan_data <- data
+  stan_data$.join_data <- NULL
   fit <- model$optimize(
-    data = data,
+    data = stan_data,
     seed = seed,
     refresh = refresh,
     init = init,
