+++ R/tar_stan_vb.R
@@ -319,8 +319,10 @@ tar_stan_vb_run <- function(
   if (is.null(seed)) {
     seed <- abs(targets::tar_seed()) + 1L
   }
+  stan_data <- data
+  stan_data$.join_data <- NULL
   fit <- model$variational(
-    data = data,
+    data = stan_data,
     seed = seed,
     refresh = refresh,
     init = init,
