+++ tests/testthat/test-tar_stan_gq_rep_summary.R
@@ -100,6 +100,10 @@ targets::tar_test("tar_stan_gq_rep_summary(compile = \"original\")", {
   expect_equal(length(table(out2$.rep)), 4L)
   expect_equal(nrow(out1), 40L)
   expect_equal(nrow(out2), 40L)
+  original_data <- tar_read(model_data)
+  y_rep <- original_data[[1]][[1]]$.join_data$y_rep
+  out1 <- out1[out1$.rep == out1$.rep[1], ]
+  expect_equal(out1$.join_data[grepl("y_rep", out1$variable)], y_rep)
   # Everything should be up to date.
   expect_equal(targets::tar_outdated(callr_function = NULL), character(0))
   # Change the model.
