+++ man/tar_stan_mcmc_rep_draws.Rd
@@ -70,8 +70,13 @@ Serves as a prefix for target names.}
 \item{stan_files}{Character vector of paths to known existing Stan model
 files created before running the pipeline.}
 
-\item{data}{Code to generate one replication of a simulated dataset.
-All models iterate over the same replicated datasets.}
+\item{data}{Code to generate a single replication of a simulated dataset.
+The workflow simulates multiple datasets, and each
+model runs on each dataset. To join data on to the model
+summaries, include a \code{.join_data}
+element of your Stan data list with names and dimensions corresponding
+to those of the model. For details, read
+\url{https://wlandau.github.io/stantargets/articles/mcmc_rep.html}.}
 
 \item{batches}{Number of batches. Each batch is a branch target
 that generates a dataset and runs the model \code{reps} times.}
@@ -293,9 +298,10 @@ default) then all variables are included.}
 
 \item{data_copy}{Character vector of names of scalars in \code{data}.
 These values will be inserted as columns in the output data frame
-for each rep. Useful for simulation studies where you want to
-check the results against some "true value" in the data. See the
-\code{mcmc_rep} vignette for an example.}
+for each rep. To join more than just scalars, include a \code{.join_data}
+element of your Stan data list with names and dimensions corresponding
+to those of the model. For details, read
+\url{https://wlandau.github.io/stantargets/articles/mcmc_rep.html}.}
 
 \item{tidy_eval}{Logical, whether to enable tidy evaluation
 when interpreting \code{command} and \code{pattern}. If \code{TRUE}, you can use the
