+++ vignettes/mcmc_rep.Rmd
@@ -32,7 +32,7 @@ It is sometimes desirable to run one or more Bayesian models repeatedly across m
 1. Validate the implementation of a Bayesian model using simulation [@cook2006; @talts2020].
 2. Simulate a randomized controlled experiment to explore frequentist properties such as power and Type I error.
 
-This vignette focuses on (1). The goal is to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated credible intervals for `beta` capture the true value of `beta` from the simulation.
+This vignette focuses on (1). The goal is to simulate multiple datasets from the model below, analyze each dataset, and assess how often the estimated credible intervals cover the true parameters from the simulation.
 
 ```{r}
 lines <- "data {
@@ -41,16 +41,19 @@ lines <- "data {
   vector[n] y;
 }
 parameters {
-  real beta;
+  real alpha;
+  vector[n] beta;
 }
 model {
-  y ~ normal(x * beta, 1);
+  y ~ normal(alpha + x .* beta, 1);
+  alpha ~ normal(0, 1);
   beta ~ normal(0, 1);
 }"
 writeLines(lines, "model.stan")
 ```
 
-Next, we define a pipeline to simulate multiple datasets and fit each dataset with the model. Below, we commit to 10 replications: 2 batches with 5 iterations per batch. (In practical situations, the total number of replications should be hundreds of times more.) We also supply custom variable names and summary functions to return the 95% credible intervals for `beta`. Below, we use the `data_copy` argument to copy the true value of `beta` from the data to the results, which will help us assess the credible intervals.
+Next, we define a pipeline to simulate multiple datasets and fit each dataset with the model. Below, we commit to 10 replications: 2 batches with 5 iterations per batch. (In practical situations, the total number of replications should be hundreds of times more.) Our input data has a special 
+`.join_data` list, which gets appended scalar-by-scalar to the numerical summaries of the correspondingly named model parameters.
 
 ```{r, echo = FALSE}
 library(targets)
@@ -58,15 +61,26 @@ tar_script({
   library(stantargets)
   options(crayon.enabled = FALSE)
   tar_option_set(memory = "transient", garbage_collection = TRUE)
+  simulate_data <- function(n = 10L) {
+    alpha <- rnorm(n = 1, mean = 0, sd = 1)
+    beta <- rnorm(n = n, mean = 0, sd = 1)
+    x <- seq(from = -1, to = 1, length.out = n)
+    y <- rnorm(n, alpha + x * beta, 1)
+    list(
+      n = n,
+      x = x,
+      y = y,
+      .join_data = list(alpha = alpha, beta = beta)
+    )
+  }
   list(
     tar_stan_mcmc_rep_summary(
       model,
       "model.stan",
-      tar_stan_example_data(),
+      simulate_data(),
       batches = 5, # Number of branch targets.
       reps = 2, # Number of model reps per branch target.
-      data_copy = "true_beta", # Append scalars from data to the output data frame.
-      variables = "beta",
+      variables = c("alpha", "beta"),
       summaries = list(
         ~posterior::quantile2(.x, probs = c(0.025, 0.975))
       ),
@@ -83,22 +97,27 @@ library(stantargets)
 options(crayon.enabled = FALSE)
 tar_option_set(memory = "transient", garbage_collection = TRUE)
 
[…]
