### 2021-03-09 04f8c8b
Remove superfluous function
