### 2021-03-10 #430 @sakrejda
@wlandau to make my life easier reviewing, is there a particular overall approach to the API, which functions you are exposing to users and which ones are considered "internal"? I don't immediately recognize the structure.

To be more specific, in the [stan compilation functions](https://github.com/wlandau/stantargets/blob/main/R/tar_stan_compile.R) there are quite a few layers of functions before you actually get down to calling the stan compilation function and it would be helpful to have an idea of why each layer is there.  Just to be clear, I don't disagree and I sometimes take a similar approach, but I'm not sure what the goals are there.  I assume this is something that happens consistently across the package.  Can you reply here pointing out the reasons for the different layers?
