### 2021-03-10 #430 @wlandau
Most `stantargets` interface functions create [target objects](https://docs.ropensci.org/targets/reference/tar_target.html). Each target object has a name, some settings, and an R command that will run later (during `tar_make()`). Putting these things together involves some layers of setup and metaprogramming. These are the layers for standalone compilation.

1. `tar_stan_compile()`: diffuse/quote any user-supplied language objects and move on to the next layer. In other words, users should be able to write `tar_stan_compile(target_name, "model_file.stan")` without needing to evaluate the symbol `target_name`. This may seem trivial in this particular example, but for other target factories like `tar_stan_mcmc()`, we also need capture the user-defined expression that generates the Stan data (and support tidy evaluation, which is best done in the first call frame).
2. `tar_stan_compile_raw()`: Gather all the settings and expressions from (1) into a target object. This layer takes a 10000-foot view of the whole target and cannot afford to go down any rabbit holes. For the sake of keeping a clean codebase and breaking down the problem into manageable pieces, it delegates the more specialized tasks to subsequent deeper layers. In `targets` and packages I build on top of `targets`, the "_raw" suffix indicates this intent.
3. `tar_stan_compile_command()`: In `stantargets`, not all the R expressions come directly from user input. Target factories in `stantargets` automatically supply many of these expressions so the user does not need to write as much custom code. The metaprogramming to create an expression from non-language input is involved enough to need its own layer.
4. `tar_stan_compile_run()`: Actually compile a Stan model when called. The "*_run()" functions significantly reduce burden of metaprogramming in layers like (3). Rather than bootstrapping a large elaborate code chunk in (3), (3) can simply create a function call to invoke (4). (4) usually has no metaprogramming, so it is a better place to manage the complicated details of how to properly invoke `cmdstanr`.

(1) is the only layer the user needs to see. (4) still needs to be exported because functions like `tar_stan_compile_run()` are part of the commands of the generated target objects, but I hide these functions from the user using `@keywords internal`.
