+++ .github/workflows/cover.yaml
@@ -11,6 +11,12 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUBPAT }}
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
     steps:
+      - name: cmdstan env vars
+        run: |
+          echo "CMDSTAN_PATH=${HOME}/.cmdstanr" >> $GITHUB_ENV
+          echo "CMDSTAN_VERSION=2.26.1" >> $GITHUB_ENV
+        shell: bash
+
       - uses: actions/checkout@v2
 
       - uses: r-lib/actions/setup-r@master
@@ -37,9 +43,18 @@ jobs:
           remotes::install_cran("covr")
         shell: Rscript {0}
 
+      - name: Cache cmdstan
+        uses: actions/cache@v2
+        with:
+          path: ${{ env.CMDSTAN_PATH }}
+          key: ${{ runner.os }}-cmdstan-${{ env.CMDSTAN_VERSION }}
+          restore-keys: ${{ runner.os }}-cmdstan-
+
       - name: Install cmdstan
         run: |
-          cmdstanr::install_cmdstan(cores = 2, overwrite = TRUE, release_url = "https://github.com/stan-dev/cmdstan/releases/download/v2.25.0/cmdstan-2.25.0.tar.gz")
+          version <- Sys.getenv("CMDSTAN_VERSION")
+          url <- sprintf("https://github.com/stan-dev/cmdstan/releases/download/v%s/cmdstan-%s.tar.gz", version, version)
+          cmdstanr::install_cmdstan(cores = 2, release_url = url)
         shell: Rscript {0}
 
       - name: Test coverage
