+++ .github/workflows/pkgdown.yaml
@@ -41,9 +41,8 @@ jobs:
 
       - name: Install dependencies
         run: |
-          install.packages("remotes")
+          install.packages(c("pkgdown", "remotes"))
           remotes::install_deps(dependencies = TRUE, type = "binary")
-          remotes::install_dev("pkgdown")
         shell: Rscript {0}
 
       - name: Cache cmdstan
