+++ .github/workflows/cover.yaml
@@ -10,11 +10,12 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUBPAT }}
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
+      CMDSTAN_VERSION: 2.26.1
+      
     steps:
       - name: cmdstan env vars
         run: |
           echo "CMDSTAN_PATH=${HOME}/.cmdstanr" >> $GITHUB_ENV
-          echo "CMDSTAN_VERSION=2.26.1" >> $GITHUB_ENV
         shell: bash
 
       - uses: actions/checkout@v2
