+++ .github/workflows/pkgdown.yaml
@@ -12,11 +12,11 @@ jobs:
     env:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: false
+      CMDSTAN_VERSION: 2.26.1
     steps:
       - name: cmdstan env vars
         run: |
           echo "CMDSTAN_PATH=${HOME}/.cmdstanr" >> $GITHUB_ENV
-          echo "CMDSTAN_VERSION=2.26.1" >> $GITHUB_ENV
         shell: bash
     
       - uses: actions/checkout@v2
