+++ .github/workflows/pkgdown.yaml
@@ -1,10 +1,5 @@
 # Some of this workflow was borrowed from https://github.com/stan-dev/cmdstanr/blob/master/.github/workflows
 # under the BSD 3-Clause License: https://github.com/stan-dev/cmdstanr/blob/master/LICENSE.md.
-on:
-  push:
-    branches: main
-# Some of this workflow was borrowed from https://github.com/stan-dev/cmdstanr/blob/master/.github/workflows
-# under the BSD 3-Clause License: https://github.com/stan-dev/cmdstanr/blob/master/LICENSE.md.
 # 
 # BSD 3-Clause License
 # Copyright (c) 2019, Stan Developers and their Assignees
@@ -35,6 +30,10 @@ on:
 # WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE 
 # OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, 
 # EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+
+on:
+  push:
+    branches: main
 
 name: pkgdown
 
