+++ .github/workflows/check.yaml
@@ -60,16 +60,13 @@ jobs:
       CMDSTAN_VERSION: 2.26.1
 
     steps:
-      - name: Non-Windows cmdstan env vars
-        if: runner.os != 'Windows'
+      - name: Platform-independent cmdstan path
+        run: writeLines(normalizePath("~/.cmdstanr"), ".github/cmdstan_path")
+        shell: Rscript {0}
+
+      - name: cmdstan env vars
         run: |
           echo "CMDSTAN_PATH=${HOME}/.cmdstanr" >> $GITHUB_ENV
-        shell: bash
-
-      - name: Windows cmdstan env vars
-        if: runner.os == 'Windows'
-        run: |
-          echo "CMDSTAN_PATH=${HOME}/Documents/.cmdstanr" >> $GITHUB_ENV
         shell: bash
     
       - uses: actions/checkout@v2
