### 2021-03-11 7de6f00
Run path stuff after setting up R
