+++ .github/workflows/check.yaml
@@ -67,7 +67,7 @@ jobs:
           r-version: ${{ matrix.config.r }}
 
       - name: Platform-independent cmdstan path
-        run: writeLines(normalizePath(file.path("~", ".cmdstanr"), winslash = "/", mustWork = FALSE), ".github/cmdstan_path")
+        run: writeLines(normalizePath(file.path("~", ".cmdstanr"), mustWork = FALSE), ".github/cmdstan_path")
         shell: Rscript {0}
 
       - name: cmdstan env vars
