+++ .github/workflows/check.yaml
@@ -67,7 +67,11 @@ jobs:
           r-version: ${{ matrix.config.r }}
 
       - name: Platform-independent cmdstan path
-        run: writeLines(normalizePath(file.path("~", ".cmdstanr"), mustWork = FALSE), ".github/cmdstan_path")
+        run: |
+          install.packages("fs")
+          path <- fs::path_rel(normalizePath(file.path("~", ".cmdstanr"), mustWork = FALSE))
+          print(path)
+          writeLines(path, ".github/cmdstan_path")
         shell: Rscript {0}
 
       - name: cmdstan env vars
