### 2021-03-12 1b7e9c2
Get absolute path
