+++ R/tar_stan_mle_rep.R
@@ -241,6 +241,7 @@ tar_stan_mle_rep <- function(
     target_lines = target_lines,
     target_data = target_data,
     target_output = target_output,
+    packages = packages,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
