+++ R/tar_stan_vb_rep.R
@@ -240,6 +240,7 @@ tar_stan_vb_rep <- function(
     target_lines = target_lines,
     target_data = target_data,
     target_output = target_output,
+    packages = packages,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
