### 2021-03-15 #430 @sakrejda
When running `devtools::test` tests fail unless `cmdstanr::install_cmdstan` is run first, I think it would be helpful to check the `cmdstan` installation first.  There is the `cmdstanr::check_cmdstan_toolchain` function that could be run in an initial test, with the `fix=TRUE` option.
