### 2021-03-15 #430 @wlandau
Good point. `keras` handles a similar situation with [`skip_if_not_tensorflow_version ()`](https://github.com/rstudio/keras/blob/master/tests/testthat/helper-skips.R), and `testthat::skip_if_not_installed()` for `Suggests:` packages is a community-accepted pattern. Personally, I think this makes sense. Do you agree?
