### 2021-03-15 #430 @sakrejda
That's a great solution, also thanks for the prompt replies!
