### 2021-03-15 981098b
Skip tests if CmdStan is not installed
