+++ tests/testthat/test-tar_stan_mle.R
@@ -2,6 +2,7 @@
 # to avoid accidentally writing to the user's file space.
 targets::tar_test("tar_stan_mle(compile = \"original\")", {
   skip_on_cran()
+  skip_if_missing_cmdstan()
   restore_compiled_models()
   targets::tar_script({
     tar_option_set(memory = "transient", garbage_collection = TRUE)
@@ -115,6 +116,7 @@ targets::tar_test("tar_stan_mle(compile = \"original\")", {
 
 targets::tar_test("tar_stan_mle(compile = \"copy\") with custom summaries", {
   skip_on_cran()
+  skip_if_missing_cmdstan()
   skip_compile_copy()
   tar_stan_example_file(path = "a.stan")
   tar_stan_example_file(path = "b.stan")
