+++ tests/testthat/test-aaa-compile.R
@@ -1,4 +1,5 @@
 tar_test("precompile models", {
+  skip_if_missing_cmdstan()
   compile_models()
   for (file in c("a.stan", "b.stan", "a", "b")) {
     expect_true(file.exists(compiled_path("a.stan")))
