### 2021-03-16 #430 @sakrejda
Aside that's really for the base `targets` pacakge, but mentioning it here since it affects running the README.md example on one of my systems:

```
> targets::tar_visnetwork()
Error in viewer(index_html) :
  'browser' must be a non-empty character string
```

From the error it's unclear that the answer is `option(browser = "<path/to/broswer>")`. It would be pretty straightforward to catch the lack of a default browser so that the message could specifically say "A default browser is required to view the network visualization, set a default browser with 'options(browser = "</path/to/browser>")`" or something similar.  From the current message I wasn't sure if I need to pass a path to the function or set an option and if I was setting an option I wasn't sure what the option was called.
