### 2021-03-16 #430 @sakrejda
In the `README.md` I initially got an error b/c `x.stan` was not found so since I was running from the package root I changed it to `inst/example.stan` and that made `tar_make()` complete successfully.  For a slightly intimidating package it would be helpful to have the default example in the README.md just work. I think you could retrieve the .stan file through a call to `system.file` so that the example runs without any changes.  It would give a new user some confidence that they could figure the package out.
