### 2021-03-16 #430 @sakrejda
The `README.md` example seems to reference an internal data generating function for the example rather than the function included in the example (not sure I got that right but I think I did).  Here's what I modified it to in order to have it working:

```# _targets.R
library(targets)
library(stantargets)

generate_data <- function(n = 10) {
  true_beta <- stats::rnorm(n = 1, mean = 0, sd = 1)
  x <- seq(from = -1, to = 1, length.out = n)
  y <- stats::rnorm(n, x * true_beta, 1)
  list(n = n, x = x, y = y, true_beta = true_beta)
}

list(
  tar_stan_mcmc(test, "inst/example.stan", generate_data())
)
```

Also I wasn't sure exactly how `stantargets` relates to `targets` and it would be helpful in the `README.md` to have a short paragraph (maybe before the "prerequisites" section?) that states that the `tar_stan_mcmc` target defines sub-targets following a specific naming convention which allow you to schedule chains and access different components of the results.  I don't want to nit-pick but taking the perspective of a naive user (I know Stan but I'm new-ish to `targets` itself) the auto-generation of a bunch of new targets is not obvious and I wasn't initially understanding that I could use `tar_read` on the auto-generated targets.
