### 2021-03-16 9b1f0a4
Use generate_data() in README
