### 2021-03-16 #430 @wlandau
> The README.md example seems to reference an internal data generating function for the example rather than the function included in the example

Thanks, just updated the README to use `generate_data()`. ([`tar_stan_example_data()`](https://wlandau.github.io/stantargets/reference/tar_stan_example_data.html) is exported, but `generate_data()` is still better since most users will write their own data functions.)

> Also I wasn't sure exactly how stantargets relates to targets and it would be helpful in the README.md to have a short paragraph (maybe before the "prerequisites" section?) that states that the tar_stan_mcmc target defines sub-targets following a specific naming convention which allow you to schedule chains and access different components of the results.

I try to address these points early on in the [first vignette](https://wlandau.github.io/stantargets/articles/mcmc.html). I would prefer to keep the README at a high level in order to keep it digestible for new users and to avoid repeating documentation.
