### 2021-03-17 8a07cb2
Implement logging in compilation targets
