+++ R/tar_stan_gq_rep_draws.R
@@ -101,7 +101,7 @@ tar_stan_gq_rep_draws <- function(
     combine = combine,
     compile = compile,
     quiet = quiet,
-    log = log,
+    log = substitute(log),
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
