+++ R/tar_stan_gq_rep_summary.R
@@ -102,7 +102,7 @@ tar_stan_gq_rep_summary <- function(
     combine = combine,
     compile = compile,
     quiet = quiet,
-    log = log,
+    log = substitute(log),
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
