+++ R/tar_stan_mcmc.R
@@ -188,7 +188,7 @@ tar_stan_mcmc <- function(
     data = sym_data,
     compile = compile,
     quiet = quiet,
-    log = log,
+    log = substitute(log),
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
