+++ R/tar_stan_vb_rep_summary.R
@@ -104,7 +104,7 @@ tar_stan_vb_rep_summary <- function(
     combine = combine,
     compile = compile,
     quiet = quiet,
-    log = log,
+    log = substitute(log),
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
