+++ R/tar_stan_mle.R
@@ -114,10 +114,6 @@ tar_stan_mle <- function(
   sym_lines <- rlang::sym(name_lines)
   sym_data <- rlang::sym(name_data)
   sym_mle <- rlang::sym(name_mle)
-  command_lines <- call_function(
-    "readLines",
-    args = list(con = rlang::sym(name_file))
-  )
   command_data <- tidy_eval(
     substitute(data),
     envir = envir,
@@ -180,7 +176,7 @@ tar_stan_mle <- function(
   )
   target_lines <- targets::tar_target_raw(
     name = name_lines,
-    command = command_lines,
+    command = command_lines(sym_file),
     packages = character(0),
     error = error,
     memory = memory,
