+++ R/tar_stan_package.R
@@ -15,6 +15,7 @@
 #' @name stantargets-package
 #' @importFrom cmdstanr cmdstan_model
 #' @importFrom digest digest
+#' @importFrom fs path_rel
 #' @importFrom fst read_fst
 #' @importFrom qs qread
 #' @importFrom posterior as_draws_df
