+++ tests/testthat/test-tar_stan_compile.R
@@ -14,5 +14,7 @@ targets::tar_test("tar_stan_compile()", {
     )
   })
   targets::tar_make(callr_function = NULL)
-  expect_equal(targets::tar_read(compile), "a.stan")
+  out <- targets::tar_read(compile)
+  expect_equal(length(out), 2)
+  expect_equal(out[1], "a.stan")
 })

[…]
