+++ man/tar_stan_mcmc_run.Rd
@@ -74,10 +74,35 @@ no longer needs access to the original Stan model file on your
 local machine. However, as a result, the Stan model re-compiles
 every time the main target reruns.}
 
+\item{quiet}{(logical) Should the verbose output from CmdStan during
+compilation be suppressed? The default is \code{TRUE}, but if you encounter an
+error we recommend trying again with \code{quiet=FALSE} to see more of the
+output.}
+
 \item{log}{Character of length 1, file path to write the stdout stream
 of the model when it runs. Set to \code{NULL} to print to the console.
 Set to \code{R.utils::nullfile()} to completely suppress all output.
 Does not apply to messages, warnings, or errors.}
+
+\item{dir}{(string) The path to the directory in which to store the CmdStan
+executable (or \code{.hpp} file if using \verb{$save_hpp_file()}). The default is the
+same location as the Stan program.}
+
+\item{include_paths}{(character vector) Paths to directories where Stan
+should look for files specified in \verb{#include} directives in the Stan
+program.}
+
+\item{cpp_options}{(list) Any makefile options to be used when compiling the
+model (\code{STAN_THREADS}, \code{STAN_MPI}, \code{STAN_OPENCL}, etc.). Anything you would
+otherwise write in the \code{make/local} file.}
+
+\item{stanc_options}{(list) Any Stan-to-C++ transpiler options to be used
+when compiling the model. See the \strong{Examples} section below as well as the
+\code{stanc} chapter of the CmdStan Guide for more details on available options:
+https://mc-stan.org/docs/cmdstan-guide/stanc.html.}
+
+\item{force_recompile}{(logical) Should the model be recompiled even if was
+not modified since last compiled. The default is \code{FALSE}.}
 
 \item{seed}{(positive integer) A seed for the (P)RNG to pass to CmdStan.}
 
