+++ NEWS.md
@@ -1,9 +1,12 @@
+# stantargets 0.0.0.9002
+
+* Return the executable file after the Stan source file in model compilation targets.
+
 # stantargets 0.0.0.9001
 
 * Join on data to summary output using `.join_data` in the Stan data (#18).
 * Pre-compile models for testing and add an environment variable to skip tests that always force recompilation (#19).
 * Load packages for any target computing summaries.
-* Return the executable file after the Stan source file in model compilation targets.
 
 # stantargets 0.0.0.9000
 
