+++ DESCRIPTION
@@ -14,7 +14,7 @@ Description: Bayesian data analysis usually incurs long runtimes
   both single-fit workflows and multi-rep simulation studies.
   For the statistical methodology, please refer to 'Stan' documentation
   (Stan Development Team 2020) <https://mc-stan.org/>.
-Version: 0.0.0.9001
+Version: 0.0.0.9002
 License: MIT + file LICENSE
 URL: https://wlandau.github.io/stantargets/, https://github.com/wlandau/stantargets
 BugReports: https://github.com/wlandau/stantargets/issues
