### 2021-03-17 68d0e77
Revert compile = "copy" tests
