+++ tests/testthat/test-tar_stan_mcmc_rep_diagnostics.R
@@ -185,10 +185,8 @@ targets::tar_test("tar_stan_mcmc_rep_diagnostics(compile = \"copy\")", {
   meta <- tar_meta(starts_with("model_data_"))
   expect_equal(nrow(meta), 2L)
   # results
-  expect_equal(length(targets::tar_read(model_file_a)), 2)
-  expect_equal(length(targets::tar_read(model_file_b)), 2)
-  expect_equal(targets::tar_read(model_file_a)[1], "a.stan")
-  expect_equal(targets::tar_read(model_file_b)[1], "b.stan")
+  expect_equal(targets::tar_read(model_file_a), "a.stan")
+  expect_equal(targets::tar_read(model_file_b), "b.stan")
   out <- targets::tar_read(model_data)
   expect_equal(length(out), 2L)
   out <- out[[2]]
