### 2021-03-18 b88adb6
Use stdout, start stderr
