+++ R/tar_stan_mcmc_rep_summary.R
@@ -39,7 +39,7 @@
 #'     data = tar_stan_example_data(),
 #'     batches = 2,
 #'     reps = 2,
-#'     log = R.utils::nullfile()
+#'     stdout = R.utils::nullfile()
 #'   )
 #' )
 #' }, ask = FALSE)
@@ -55,7 +55,7 @@ tar_stan_mcmc_rep_summary <- function(
   combine = TRUE,
   compile = c("original", "copy"),
   quiet = TRUE,
-  log = NULL,
+  stdout = NULL,
   dir = NULL,
   include_paths = NULL,
   cpp_options = list(),
@@ -115,7 +115,7 @@ tar_stan_mcmc_rep_summary <- function(
     combine = combine,
     compile = match.arg(compile),
     quiet = quiet,
-    log = substitute(log),
+    stdout = stdout,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
