+++ R/tar_stan_vb_rep.R
@@ -35,7 +35,7 @@ tar_stan_vb_rep <- function(
   combine = TRUE,
   compile = c("original", "copy"),
   quiet = TRUE,
-  log = NULL,
+  stdout = NULL,
   dir = NULL,
   include_paths = NULL,
   cpp_options = list(),
@@ -107,7 +107,7 @@ tar_stan_vb_rep <- function(
     output = match.arg(output),
     compile = compile,
     quiet = quiet,
-    log = log,
+    stdout = stdout,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
@@ -153,7 +153,7 @@ tar_stan_vb_rep <- function(
     name = name_file,
     stan_file = quote(._stantargets_file_50e43091),
     quiet = quiet,
-    log = log,
+    stdout = stdout,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
@@ -265,7 +265,7 @@ tar_stan_vb_rep_run <- function(
   output,
   compile,
   quiet,
-  log,
+  stdout,
   dir,
   include_paths,
   cpp_options,
@@ -292,9 +292,13 @@ tar_stan_vb_rep_run <- function(
   summaries,
   summary_args
 ) {
-  if (!is.null(log)) {
-    sink(file = log, type = "output", append = TRUE)
+  if (!is.null(stdout)) {
+    sink(file = stdout, type = "output", append = TRUE)
     on.exit(sink(file = NULL, type = "output"))
+  }
+  if (!is.null(stderr)) {
+    sink(file(stderr, "at"), type = "message", append = TRUE)
+    on.exit(sink(file = NULL, type = "message"))
   }
   file <- stan_file
   if (identical(compile, "copy")) {
