+++ tests/testthat/test-tar_stan_mcmc_rep_draws.R
@@ -23,7 +23,7 @@ targets::tar_test("tar_stan_mcmc_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        log = R.utils::nullfile()
+        stdout = R.utils::nullfile()
       )
     )
   })
@@ -116,7 +116,7 @@ targets::tar_test("tar_stan_mcmc_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        log = R.utils::nullfile()
+        stdout = R.utils::nullfile()
       )
     )
   })
@@ -152,7 +152,7 @@ targets::tar_test("tar_stan_mcmc_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        log = R.utils::nullfile()
+        stdout = R.utils::nullfile()
       )
     )
   })
@@ -245,7 +245,7 @@ targets::tar_test("tar_stan_mcmc_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        log = R.utils::nullfile()
+        stdout = R.utils::nullfile()
       )
     )
   })
