+++ R/tar_stan_vb_rep_draws.R
@@ -40,7 +40,8 @@
 #'     data = tar_stan_example_data(),
 #'     batches = 2,
 #'     reps = 2,
-#'     stdout = R.utils::nullfile()
+#'     stdout = R.utils::nullfile(),
+#'     stderr = R.utils::nullfile()
 #'   )
 #' )
 #' }, ask = FALSE)
@@ -57,6 +58,7 @@ tar_stan_vb_rep_draws <- function(
   compile = c("original", "copy"),
   quiet = TRUE,
   stdout = NULL,
+  stderr = NULL,
   dir = NULL,
   include_paths = NULL,
   cpp_options = list(),
@@ -104,6 +106,7 @@ tar_stan_vb_rep_draws <- function(
     compile = compile,
     quiet = quiet,
     stdout = stdout,
+    stderr = stderr,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
