+++ tests/testthat/test-tar_stan_mle_rep_draws.R
@@ -18,7 +18,8 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -47,7 +48,7 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"original\")", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   meta <- tar_meta(starts_with("model_data_"))
   expect_equal(nrow(meta), 2L)
   expect_equal(length(targets::tar_read(model_file_x)), 2)
@@ -106,7 +107,8 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -137,7 +139,8 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -168,7 +171,7 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"copy\") custom", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   meta <- tar_meta(starts_with("model_data_"))
   expect_equal(nrow(meta), 2L)
   expect_equal(targets::tar_read(model_file_a), "a.stan")
@@ -221,7 +224,8 @@ targets::tar_test("tar_stan_mle_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
