+++ R/tar_stan_gq_rep.R
@@ -38,6 +38,7 @@ tar_stan_gq_rep <- function(
   compile = c("original", "copy"),
   quiet = TRUE,
   stdout = NULL,
+  stderr = NULL,
   dir = NULL,
   include_paths = NULL,
   cpp_options = list(),
@@ -100,6 +101,7 @@ tar_stan_gq_rep <- function(
     compile = compile,
     quiet = quiet,
     stdout = stdout,
+    stderr = stderr,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
@@ -135,6 +137,7 @@ tar_stan_gq_rep <- function(
     stan_file = quote(._stantargets_file_50e43091),
     quiet = quiet,
     stdout = stdout,
+    stderr = stderr,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
@@ -248,6 +251,7 @@ tar_stan_gq_rep_run <- function(
   compile,
   quiet,
   stdout,
+  stderr,
   dir,
   include_paths,
   cpp_options,
@@ -268,8 +272,12 @@ tar_stan_gq_rep_run <- function(
     on.exit(sink(file = NULL, type = "output"))
   }
   if (!is.null(stderr)) {
-    sink(file(stderr, "at"), type = "message", append = TRUE)
-    on.exit(sink(file = NULL, type = "message"))
+    con <- file(stderr, "at")
+    sink(con, type = "message", append = TRUE)
+    on.exit({
+      sink(file = NULL, type = "message")
+      close(con)
+    }, add = TRUE)
   }
   file <- stan_file
   if (identical(compile, "copy")) {
