+++ R/tar_stan_mcmc_rep_diagnostics.R
@@ -44,7 +44,8 @@
 #'     data = tar_stan_example_data(),
 #'     batches = 2,
 #'     reps = 2,
-#'     stdout = R.utils::nullfile()
+#'     stdout = R.utils::nullfile(),
+#'     stderr = R.utils::nullfile()
 #'   )
 #' )
 #' }, ask = FALSE)
@@ -61,6 +62,7 @@ tar_stan_mcmc_rep_diagnostics <- function(
   compile = c("original", "copy"),
   quiet = TRUE,
   stdout = NULL,
+  stderr = NULL,
   dir = NULL,
   include_paths = NULL,
   cpp_options = list(),
@@ -119,6 +121,7 @@ tar_stan_mcmc_rep_diagnostics <- function(
     compile = match.arg(compile),
     quiet = quiet,
     stdout = stdout,
+    stderr = stderr,
     dir = dir,
     include_paths = include_paths,
     cpp_options = cpp_options,
