+++ tests/testthat/test-tar_stan_vb_rep_draws.R
@@ -17,7 +17,8 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -46,7 +47,7 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"original\")", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   meta <- tar_meta(starts_with("model_data_"))
   expect_equal(nrow(meta), 2L)
   expect_equal(length(targets::tar_read(model_file_x)), 2)
@@ -106,7 +107,8 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"original\")", {
         batches = 2,
         reps = 2,
         combine = TRUE,
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -138,7 +140,8 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })
@@ -169,7 +172,7 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"copy\") custom", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   meta <- tar_meta(starts_with("model_data_"))
   expect_equal(nrow(meta), 2L)
   expect_equal(targets::tar_read(model_file_a), "a.stan")
@@ -227,7 +230,8 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"copy\") custom", {
         combine = TRUE,
         data_copy = c("n", "true_beta"),
         variables = "beta",
-        stdout = R.utils::nullfile()
+        stdout = R.utils::nullfile(),
+        stderr = R.utils::nullfile()
       )
     )
   })

[…]
