+++ R/tar_stan_mcmc_rep.R
@@ -360,16 +360,10 @@ tar_stan_mcmc_rep_run <- function(
   summary_args
 ) {
   if (!is.null(stdout)) {
-    sink(file = stdout, type = "output", append = TRUE)
-    on.exit(sink(file = NULL, type = "output"))
+    withr::local_output_sink(new = stdout, append = TRUE, split = TRUE)
   }
   if (!is.null(stderr)) {
-    con <- file(stderr, "at")
-    sink(con, type = "message", append = TRUE)
-    on.exit({
-      sink(file = NULL, type = "message")
-      close(con)
-    }, add = TRUE)
+    withr::local_message_sink(new = stderr, append = TRUE)
   }
   file <- stan_file
   if (identical(compile, "copy")) {
