+++ DESCRIPTION
@@ -45,15 +45,15 @@ Imports:
   targets (>= 0.0.1),
   tarchetypes (>= 0.0.1),
   tibble (>= 3.0.4),
-  tools
+  tools,
+  withr (>= 2.1.2)
 Suggests:
   dplyr (>= 1.0.2),
   knitr (>= 1.28),
   R.utils (>= 2.10.1),
   rmarkdown (>= 2.1),
   testthat (>= 3.0.0),
-  visNetwork (>= 2.0.9),
-  withr (>= 2.1.2)
+  visNetwork (>= 2.0.9)
 Remotes:
   stan-dev/cmdstanr,
   stan-dev/posterior
