+++ tests/testthat/test-tar_stan_mcmc.R
@@ -49,7 +49,7 @@ targets::tar_test("tar_stan_mcmc(compile = \"original\")", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  suppressWarnings(targets::tar_make(callr_function = NULL))
+  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
   expect_equal(targets::tar_read(model_file_x), "a.stan")
   expect_equal(targets::tar_read(model_file_y), "b.stan")
   out <- targets::tar_read(model_data)
@@ -193,7 +193,7 @@ targets::tar_test("tar_stan_mcmc(compile = \"copy\") with custom summaries", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  suppressWarnings(targets::tar_make(callr_function = NULL))
+  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
   expect_equal(targets::tar_read(model_file_a), "a.stan")
   expect_equal(targets::tar_read(model_file_b), "b.stan")
   expect_equal(
