+++ R/tar_stan_mcmc_rep.R
@@ -360,7 +360,7 @@ tar_stan_mcmc_rep_run <- function(
   summary_args
 ) {
   if (!is.null(stdout)) {
-    withr::local_output_sink(new = stdout, append = TRUE, split = TRUE)
+    withr::local_output_sink(new = stdout, append = TRUE)
   }
   if (!is.null(stderr)) {
     withr::local_message_sink(new = stderr, append = TRUE)
