+++ tests/testthat/test-tar_stan_gq.R
@@ -64,7 +64,7 @@ targets::tar_test("tar_stan_gq(compile = \"original\")", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   expect_equal(targets::tar_read(gq_file_a), "a.stan")
   expect_equal(targets::tar_read(gq_file_b), "b.stan")
   out <- targets::tar_read(gq_data)
@@ -226,7 +226,7 @@ targets::tar_test("tar_stan_gq(compile = \"copy\") with custom summaries", {
   rownames(exp) <- NULL
   expect_equal(out, exp)
   # results
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   expect_equal(targets::tar_read(gq_file_a), "a.stan")
   expect_equal(targets::tar_read(gq_file_b), "b.stan")
   expect_equal(
