+++ tests/testthat/test-tar_stan_summary.R
@@ -21,7 +21,7 @@ targets::tar_test("tar_stan_summary() with defaults", {
       tar_stan_summary(summary, fit = model_mcmc_a)
     )
   })
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   out <- tar_read(summary)
   expect_true(tibble::is_tibble(out))
   expect_true(nrow(out) > 1L)
@@ -59,7 +59,7 @@ targets::tar_test("tar_stan_summary() with custom summaries", {
       )
     )
   })
-  capture.output(suppressWarnings(targets::tar_make(callr_function = NULL)))
+  suppressWarnings(targets::tar_make(callr_function = NULL))
   out <- tar_read(summary)
   expect_true(tibble::is_tibble(out))
   expect_equal(nrow(out), 1L)
