+++ R/tar_stan_gq_rep.R
@@ -268,7 +268,7 @@ tar_stan_gq_rep_run <- function(
   summary_args
 ) {
   if (!is.null(stdout)) {
-    withr::local_output_sink(new = stdout, append = TRUE, split = TRUE)
+    withr::local_output_sink(new = stdout, append = TRUE)
   }
   if (!is.null(stderr)) {
     withr::local_message_sink(new = stderr, append = TRUE)
