+++ NAMESPACE
@@ -51,5 +51,5 @@ importFrom(targets,tar_script)
 importFrom(targets,tar_target)
 importFrom(targets,tar_target_raw)
 importFrom(targets,tar_test)
-importFrom(withr,with_message_sink)
-importFrom(withr,with_output_sink)
+importFrom(withr,local_message_sink)
+importFrom(withr,local_output_sink)
