+++ man/tar_stan_mcmc_rep_draws.Rd
@@ -79,8 +79,9 @@ element of your Stan data list with names and dimensions corresponding
 to those of the model. For details, read
 \url{https://wlandau.github.io/stantargets/articles/mcmc_rep.html}.}
 
-\item{batches}{Number of batches. Each batch is a branch target
-that generates a dataset and runs the model \code{reps} times.}
+\item{batches}{Number of batches. Each batch is a sequence
+of branch targets containing multiple reps. Each rep
+generates a dataset and runs the model on it.}
 
 \item{reps}{Number of replications per batch.}
 
