+++ man/tar_stan_gq_rep_summary.Rd
@@ -70,8 +70,9 @@ VB) object returned by CmdStanR's \code{\link[cmdstanr:fit-method-draws]{$draws(
 \item A character vector of paths to CmdStan CSV output files.
 }}
 
-\item{batches}{Number of batches. Each batch is a branch target
-that generates a dataset and runs the model \code{reps} times.}
+\item{batches}{Number of batches. Each batch is a sequence
+of branch targets containing multiple reps. Each rep
+generates a dataset and runs the model on it.}
 
 \item{reps}{Number of replications per batch.}
 
