### 2021-03-24 6a1f207
Switch meanings of logical infix ops to conform to precedent
