+++ tests/testthat/test-utils_logic.R
@@ -1,17 +1,17 @@
 # targets::tar_test() runs the test code inside a temporary directory
 # to avoid accidentally writing to the user's file space.
-targets::tar_test("%||%", {
-  expect_equal("x" %||% "y", "x")
-  expect_equal(character(0) %||% "y", character(0))
-  expect_equal(NA_character_ %||% "y", NA_character_)
-  expect_equal(NULL %||% "y", "y")
-})
-
 targets::tar_test("%|||%", {
   expect_equal("x" %|||% "y", "x")
-  expect_equal(character(0) %|||% "y", "y")
+  expect_equal(character(0) %|||% "y", character(0))
   expect_equal(NA_character_ %|||% "y", NA_character_)
   expect_equal(NULL %|||% "y", "y")
+})
+
+targets::tar_test("%||%", {
+  expect_equal("x" %||% "y", "x")
+  expect_equal(character(0) %||% "y", "y")
+  expect_equal(NA_character_ %||% "y", NA_character_)
+  expect_equal(NULL %||% "y", "y")
 })
 
 targets::tar_test("%||NA%", {
