### 2021-03-24 #430 @sakrejda
There's no way to pass additional arguments here: https://github.com/wlandau/stantargets/blob/522a94c3b0941aba10d49bfc756b2a6def1a5577/R/tar_stan_compile.R#L155

That means the `pedantic` and `threads` arguments currently can't be passed down.  What do you think about adding a `compile_dot_args` to the `tar_stan_compile` function that would get passed down to the `cmdstanr::cmdstan_model` function?
