### 2021-03-24 #430 @sakrejda
I guess that's just a maintenance issue that will come up repeatedly with the way that arguments are currently structured (explicitly passed down and enumerated at each level). In some ways it's helpful, but it also means a user can end up stuck.   What do you think about this issue? Is it an explicit tradeoff you're making?
