### 2021-03-25 #430 @wlandau
`pedantic` may have been created in `cmdstanr` after I last updated `tar_stan_compile()`, and `threads` is deprecated in favor of `cpp_options`. I intend to include all the formal arguments in the interface so users have a clearer idea how to use the functions. I think this is worth the maintenance cost. I will add `pedantic`.
