### 2021-03-25 cf88558
Add pedantic (https://github.com/ropensci/software-review/issues/430#issuecomment-805987871)
