+++ R/tar_stan_gq_rep_draws.R
@@ -69,6 +69,7 @@ tar_stan_gq_rep_draws <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -107,6 +108,7 @@ tar_stan_gq_rep_draws <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
