+++ R/tar_stan_mle_rep_summary.R
@@ -59,6 +59,7 @@ tar_stan_mle_rep_summary <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -108,6 +109,7 @@ tar_stan_mle_rep_summary <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
