+++ R/tar_stan_compile.R
@@ -35,6 +35,7 @@ tar_stan_compile <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -57,6 +58,7 @@ tar_stan_compile <- function(
     stdout = stdout,
     stderr = substitute(stderr),
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -80,6 +82,7 @@ tar_stan_compile_raw <- function(
   stdout,
   stderr,
   dir,
+  pedantic,
   include_paths,
   cpp_options,
   stanc_options,
@@ -100,6 +103,7 @@ tar_stan_compile_raw <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -127,6 +131,7 @@ tar_stan_compile_command <- function(
   stdout,
   stderr,
   dir,
+  pedantic,
   include_paths,
   cpp_options,
   stanc_options,
@@ -139,6 +144,7 @@ tar_stan_compile_command <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -168,6 +174,7 @@ tar_stan_compile_run <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -185,6 +192,7 @@ tar_stan_compile_run <- function(
     compile = TRUE,
     quiet = quiet,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
