+++ R/tar_stan_vb.R
@@ -63,6 +63,7 @@ tar_stan_vb <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -143,6 +144,7 @@ tar_stan_vb <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -274,6 +276,7 @@ tar_stan_vb_run <- function(
   stdout,
   stderr,
   dir,
+  pedantic,
   include_paths,
   cpp_options,
   stanc_options,
@@ -313,6 +316,7 @@ tar_stan_vb_run <- function(
     compile = TRUE,
     quiet = quiet,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
