+++ R/tar_stan_mcmc_rep_summary.R
@@ -59,6 +59,7 @@ tar_stan_mcmc_rep_summary <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -120,6 +121,7 @@ tar_stan_mcmc_rep_summary <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
