+++ R/tar_stan_gq_rep.R
@@ -40,6 +40,7 @@ tar_stan_gq_rep <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -103,6 +104,7 @@ tar_stan_gq_rep <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -139,6 +141,7 @@ tar_stan_gq_rep <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -253,6 +256,7 @@ tar_stan_gq_rep_run <- function(
   stdout,
   stderr,
   dir,
+  pedantic,
   include_paths,
   cpp_options,
   stanc_options,
@@ -285,6 +289,7 @@ tar_stan_gq_rep_run <- function(
     compile = TRUE,
     quiet = quiet,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
