+++ R/tar_stan_mle_rep_draws.R
@@ -59,6 +59,7 @@ tar_stan_mle_rep_draws <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -106,6 +107,7 @@ tar_stan_mle_rep_draws <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
