+++ R/tar_stan_vb_rep.R
@@ -38,6 +38,7 @@ tar_stan_vb_rep <- function(
   stdout = NULL,
   stderr = NULL,
   dir = NULL,
+  pedantic = FALSE,
   include_paths = NULL,
   cpp_options = list(),
   stanc_options = list(),
@@ -111,6 +112,7 @@ tar_stan_vb_rep <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -158,6 +160,7 @@ tar_stan_vb_rep <- function(
     stdout = stdout,
     stderr = stderr,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,
@@ -271,6 +274,7 @@ tar_stan_vb_rep_run <- function(
   stdout,
   stderr,
   dir,
+  pedantic,
   include_paths,
   cpp_options,
   stanc_options,
@@ -314,6 +318,7 @@ tar_stan_vb_rep_run <- function(
     compile = TRUE,
     quiet = quiet,
     dir = dir,
+    pedantic = pedantic,
     include_paths = include_paths,
     cpp_options = cpp_options,
     stanc_options = stanc_options,

[…]
