### 2021-03-25 #430 @sakrejda
Keeping the `stantargets` functions updated to the `cmdstanr` functions makes a lot of sense for generating good user documentation so it's hard to argue with that.  I initially thought an automated tool could help (comparing formals, etc..) but the `cmdstanr` package is written with the opposite strategy where almost everything is passed down through `...` so there's no easy way to implement that.
