### 2021-03-25 #430 @sakrejda
In many places I'm noticing a mixed usage of base R's `substitute` and `rlang`'s `tidy_eval` and related functions.  Sections 19.3.3 and 19.3.4 of [Wickham's advanced R book](https://adv-r.hadley.nz/quasiquotation.html) compare `substitute` with a couple of `rlang` functions and note some differences in behavior.

1. The mix of usage and behavior in `stantargets` may end up being confusing for the user (and other maintainers).
2. There is a related [set of utility functions in need of documentation](https://github.com/wlandau/stantargets/blob/main/R/utils_language.R):
3. Those functions re-define some well-known names ('tidy_eval' for one) in a way that's not immediately obvious, is there a specific reason for that or could the names be changed?

I'm not sure what the best solution is. Ideally I would like to see a package like this stick to purely using `rlang` because that seems like the idiom that will stick around for a while, the behavior of those functions is consistent, and more users will ultimately be familiar with it.  Additionally the in-between state seems like a significant maintenance burden.. There are lots of solutions (e.g.-just documenting the specific strategy used in the package and the utility functions) so I'd like to hear what your take is.
