### 2021-03-25 #430 @sakrejda
Minor comments:

Utility function without documentation, relies on file.exists which means [`assert_path`](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/utils_assert.R#L25) will return without error even if the path is a directory, is that intentional?

The `assert_stan_file` function will return without error even if the path to the Stan file is a [path to a directory](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/utils_assert.R#L46), most users wouldn't create `<blah.stan>` directory but with automation it certainly happens.  My preferred solution is to use `fs::is_file` which wraps both the existence and type check, but `isFALSE(file.info('<path>')$isdi)r` returns the same information.

The `assert_package` function uses `requireNamespace` which loads the package namespace as a side-effect... that's probably fine although using `installed.packages` would avoid that side-effect.

For [`utils/condition.R` ](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/utils_condition.R#L2)relying on `rlang::abort` rather than `stop` would make it possible to return the objects used in whatever function failed the assertion rather than just a text string.  I realize actually making that switch might be a fairly extensive set of changes so just mentioning it.

Confusing code in [`utils/output.R`](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/utils_output.R#L21) since `output` is not defined as a parameter or locally

Stan fit objects can get very large, is there a specific reason for loading all the cached data into the object [here](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/tar_stan_mcmc.R#L457)? That approach can make it impossible to use the function due to RAM constraints.
