### 2021-03-25 0c58a0e
Use rlang::check_installed() and check that stan_file is not a directory
