+++ R/tar_stan_package.R
@@ -20,7 +20,7 @@
 #' @importFrom qs qread
 #' @importFrom posterior as_draws_df
 #' @importFrom purrr map
-#' @importFrom rlang as_function sym
+#' @importFrom rlang as_function check_installed expr sym
 #' @importFrom stats rnorm runif
 #' @importFrom targets tar_cue tar_dir tar_load tar_option_get tar_path
 #'   tar_read tar_script tar_target tar_target_raw tar_test
