+++ R/utils_assert.R
@@ -4,6 +4,12 @@ assert_chr <- function(x, msg = NULL) {
   }
 }
 
+assert_nonempty <- function(x, msg = NULL) {
+  if (!length(x)) {
+    throw_validate(msg %|||% "x must not be empty")
+  }
+}
+
 assert_not_in <- function(x, choices, msg = NULL) {
   if (any(x %in% choices)) {
     throw_validate(msg %|||% paste(deparse(x), "is in", deparse(choices)))
@@ -17,17 +23,18 @@ assert_nzchar <- function(x, msg = NULL) {
 }
 
 assert_package <- function(package, msg = NULL) {
-  if (!requireNamespace(package, quietly = TRUE)) {
-    throw_validate(msg %|||% paste("package ", package, " not installed"))
-  }
+  tryCatch(rlang::check_installed(package), error = function(e) {
+    throw_validate(conditionMessage(e))
+  })
 }
 
 assert_path <- function(path, msg = NULL) {
+  assert_nonempty(path, "path must not be empty")
   missing <- !file.exists(path)
   if (any(missing)) {
     throw_validate(
       msg %|||% paste(
-        "missing files: ",
+        "missing files or directories: ",
         paste(path[missing], collapse = ", ")
       )
     )
@@ -44,6 +51,9 @@ assert_stan_file <- function(stan_file) {
   assert_scalar(stan_file, "only one Stan model file allowed at a time.")
   assert_chr(stan_file, "stan_file must be a character.")
   assert_path(stan_file, "stan_file is not an existing file.")
+  if (dir.exists(stan_file)) {
+    throw_validate("stan_file must not be a directory.")
+  }
 }
 
 assert_unique <- function(x, msg = NULL) {
