+++ NAMESPACE
@@ -36,6 +36,8 @@ importFrom(posterior,as_draws_df)
 importFrom(purrr,map)
 importFrom(qs,qread)
 importFrom(rlang,as_function)
+importFrom(rlang,check_installed)
+importFrom(rlang,expr)
 importFrom(rlang,sym)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
