+++ R/tar_stan_gq.R
@@ -115,11 +115,11 @@ tar_stan_gq <- function(
   name_gq <- paste0(name, "_gq")
   name_draws <- paste0(name, "_draws")
   name_summary <- paste0(name, "_summary")
-  sym_stan <- rlang::syms(name_stan)
-  sym_file <- rlang::sym(name_file)
-  sym_lines <- rlang::sym(name_lines)
-  sym_data <- rlang::sym(name_data)
-  sym_gq <- rlang::sym(name_gq)
+  sym_stan <- as_symbols(name_stan)
+  sym_file <- as.symbol(name_file)
+  sym_lines <- as.symbol(name_lines)
+  sym_data <- as.symbol(name_data)
+  sym_gq <- as.symbol(name_gq)
   command_data <- tidy_eval(
     substitute(data),
     envir = envir,
