### 2021-03-25 #430 @wlandau
Re https://github.com/ropensci/software-review/issues/430#issuecomment-806590196,

> 1. The mix of usage and behavior in stantargets may end up being confusing for the user (and other maintainers).

My strategy for `rlang` is to only choose components with significant advantages over base R for this specific set of use cases. I just refactored `stantargets` a little to make that clearer. As of just now, the only functions imported from `rlang` are:

* `check_installed()`: just switched to this one because it prompts the user to install any missing packages which is much nicer than simply checking `requireNamespace()`.
* `expr()` and `quo_squash()`: `stantargets` needs to apply tidy evaluation to resolve `!!` in the commands, but without actually evaluating the rest of the command. That is the purpose of the custom [`tidy_eval()`](https://github.com/wlandau/stantargets/blob/5b4cd0df1a9c351ce5b15a766dc7164819cf1738/R/utils_language.R#L38) function.

I hesitate to depend on more `rlang` functions that already have equivalents in base R or `stantargets` because I have noticed changes in `rlang` over the years, and adjustment is a maintenance cost. Similar maintenance costs are already high because `cmdstanr` is developing rapidly.

> There is a related set of utility functions in need of documentation.

What sort of documentation do you have in mind? I do not usually write docstrings for unexported functions.

> Those functions re-define some well-known names ('tidy_eval' for one) in a way that's not immediately obvious, is there a specific reason for that or could the names be changed?

There exists an `rlang::eval_tidy()` but no `rlang::tidy_eval()`, so I thought this was safe. What other conflicts do you see?
