+++ R/tar_stan_mcmc_rep.R
@@ -46,7 +46,7 @@ tar_stan_mcmc_rep <- function(
   name,
   stan_files,
   data = list(),
-  output = c("summary", "draws", "diagnostics"),
+  output_type = c("summary", "draws", "diagnostics"),
   batches = 1L,
   reps = 1L,
   combine = TRUE,
@@ -136,7 +136,7 @@ tar_stan_mcmc_rep <- function(
     stan_name = quote(._stantargets_name_chr_50e43091),
     stan_path = quote(._stantargets_file_50e43091),
     data = sym_data,
-    output = match.arg(output),
+    output_type = match.arg(output_type),
     compile = compile,
     quiet = quiet,
     stdout = stdout,
@@ -307,7 +307,7 @@ tar_stan_mcmc_rep <- function(
 #' @inheritParams tar_stan_mcmc
 #' @param stan_name Friendly suffix of the Stan model target.
 #' @param stan_path Original path to the input Stan file.
-#' @param output Type of output to create, either `"summaries"`,
+#' @param output_type Type of output to create, either `"summaries"`,
 #'   `"draws"`, or `"diagnostics"`.
 #' @param data_copy Character vector of names of scalars in `data`.
 #'   These values will be inserted as columns in the output data frame
@@ -320,7 +320,7 @@ tar_stan_mcmc_rep_run <- function(
   stan_name,
   stan_path,
   data,
-  output,
+  output_type,
   compile,
   quiet,
   stdout,
@@ -398,7 +398,7 @@ tar_stan_mcmc_rep_run <- function(
     ~tar_stan_mcmc_rep_run_rep(
       data = .x,
       seed = .y,
-      output = output,
+      output_type = output_type,
       model = model,
       refresh = refresh,
       init = init,
@@ -440,7 +440,7 @@ tar_stan_mcmc_rep_run <- function(
 
 tar_stan_mcmc_rep_run_rep <- function(
   data,
-  output,
+  output_type,
   seed,
   model,
   refresh,
@@ -509,7 +509,7 @@ tar_stan_mcmc_rep_run_rep <- function(
   )
   tar_stan_output(
     fit = fit,
-    output = output,
+    output_type = output_type,
     summaries = summaries,
     summary_args = summary_args,
     variables = variables,
