+++ R/tar_stan_mcmc_rep_draws.R
@@ -112,7 +112,7 @@ tar_stan_mcmc_rep_draws <- function(
     name = deparse_language(substitute(name)),
     stan_files = stan_files,
     data = substitute(data),
-    output = "draws",
+    output_type = "draws",
     batches = batches,
     reps = reps,
     combine = combine,
