### 2021-03-25 #430 @wlandau
> Utility function without documentation, relies on file.exists which means assert_path will return without error even if the path is a directory, is that intentional?

Yes, I would rather `assert_path()` be narrow in scope. Depending on where else these assertions become necessary, I may write an `assert_files()` or `assert_not_dirs()` function.

> The assert_stan_file function will return without error even if the path to the Stan file is a path to a directory,

I just added a statement in `assert_stan_file()` to throw an informative error if `stan_file` is not a directory.

> The assert_package function uses requireNamespace which loads the package namespace as a side-effect... that's probably fine although using installed.packages would avoid that side-effect.

I actually just switched to `rlang::check_installed()` because it has a nice prompt for the user to install missing packages in interactive mode. This function also uses `requireNamespace()`. I am not sure there is another way to efficiently check for an installed package. `installed.packages()` is pretty slow.

> For utils/condition.R relying on rlang::abort rather than stop would make it possible to return the objects used in whatever function failed the assertion rather than just a text string. I realize actually making that switch might be a fairly extensive set of changes so just mentioning it.

Yeah, I would have definitely used `rlang::abort()` and `rlang::warn()` if I had known about them from the beginning.

> Confusing code in utils/output.R since output is not defined as a parameter or locally

I thought I defined `output` as a parameter [here](https://github.com/wlandau/stantargets/blob/cf8855837623504d5493854f303200f0e4258c6a/R/utils_output.R#L13). Just changed the name to `output_type` to make the meaning clearer.

> Stan fit objects can get very large, is there a specific reason for loading all the cached data into the object here? That approach can make it impossible to use the function due to RAM constraints.

It allows `tar_stan_mcmc()` to define downstream compressed targets for multiple different kinds of output (draws, diagnostics summaries). If we do not save the entire CmdStanMCMC object with all its data, the information vanishes by default. The alternative is to save the CSV files to persistent storage, tell `targets` to track those output files, and then bootstrap the CmdStanMCMC object using https://github.com/stan-dev/cmdstanr/pull/412 (which was not available when I first wrote `tar_stan_mcmc()`). But this requires designing a data storage system outside the `targets` data store (`_targets/`), and those CSV files are uncompressed and could take up a lot of space on disk. (`stantargets` compresses CmdStanMCMC objects with `qs` and data frames with `fst`.) Users who hit memory constraints can currently run `tar_stan_mcmc_rep_summary()` with the default settings of `batches = 1` and `data = 1`.
