### 2021-03-25 #430 @sakrejda
I noticed the changes you made in your usage of `rlang` and they're helpful.  I'm hopeful `rlang` is more stable than it used to be but it makes sense to be cautious about it.  I also like that you've been working with the `posterior` package rather than re-inventing the wheel on MCMC summaries.

I mis-read `tidy_eval` as `eval_tidy` a few times in a row before you pointed out they're different.  Not everyone reliably recognizes word order (me included) so a name like `tar_eval` would make it clear this is a function your package defines rather than something imported.

I'll have some more comments but overall I think this package fills an important niche really well and you're addressing the challenges that are going to come up in terms of maintenance.  The evolution of `cmdstanr` is probably going to keep going but I don't think you'll see huge breaking changes in terms of the arguments you end up passing down so the use of formal arguments seems like a toss-up in terms of costs and benefits.
