+++ R/tar_stan_mle.R
@@ -117,7 +117,7 @@ tar_stan_mle <- function(
   sym_lines <- as.symbol(name_lines)
   sym_data <- as.symbol(name_data)
   sym_mle <- as.symbol(name_mle)
-  command_data <- tidy_eval(
+  command_data <- tar_tidy_eval(
     substitute(data),
     envir = envir,
     tidy_eval = tidy_eval
