+++ R/tar_stan_gq.R
@@ -120,7 +120,7 @@ tar_stan_gq <- function(
   sym_lines <- as.symbol(name_lines)
   sym_data <- as.symbol(name_data)
   sym_gq <- as.symbol(name_gq)
-  command_data <- tidy_eval(
+  command_data <- tar_tidy_eval(
     substitute(data),
     envir = envir,
     tidy_eval = tidy_eval
