+++ R/tar_stan_vb_rep.R
@@ -91,7 +91,7 @@ tar_stan_vb_rep <- function(
   sym_batch <- as.symbol(name_batch)
   sym_data <- as.symbol(name_data)
   command_batch <- substitute(seq_len(x), env = list(x = batches))
-  command_rep <- tidy_eval(
+  command_rep <- tar_tidy_eval(
     data,
     envir = envir,
     tidy_eval = tidy_eval
