### 2021-03-25 #430 @wlandau
Thanks, @sakrejda.

> a name like tar_eval would make it clear this is a function your package defines rather than something imported.

Now changed to `tar_tidy_eval()`.
