+++ R/utils_condition.R
@@ -1,13 +1,13 @@
 throw_validate <- function(...) {
-  stop(condition_validate(...))
+  stop(tar_condition_validate(...))
 }
 
-condition_validate <- function(...) {
+tar_condition_validate <- function(...) {
   structure(
     list(message = paste0(..., collapse = ""), call = NULL),
     class = c(
-      "condition_validate",
-      "condition_tarchetypes",
+      "tar_condition_validate",
+      "tar_condition_tarchetypes",
       "error",
       "condition"
     )
