### 2021-03-31 #430 @mattwarkentin
@melvidoni and @wlandau, please see the attached review. This is my first time participating in software peer-review, so I've tried to follow the guidelines as much as possible. Some of my comments may or may not be out-of-scope for this kind of review, or too pedantic. I have reviewed this package primarily for following best practices for `R` package development, package usability, and for how it fits/works with `targets`, and not so much as a Bayesian, `stan`, or `cmdstanr` expert (which I believe is covered by @sakrejda). I hope this review is helpful. Thanks for including me in the process.

- I am listed as a formal contributor  the `targets` package (https://github.com/ropensci/targets/blob/main/DESCRIPTION), which was authored by the author of the current package under review (`stantargets`). For this reason, I declined to review `targets` due to a perceived COI. However, I do not believe these previous (minor) contributions represent a COI for reviewing the current package.

#### Documentation:

- [ ] **Examples** (that run successfully locally) for all exported functions

---

- As expected, this package is extremely well-developed and is yet another amazing contribution to the growing `targets` ecosystem. The code is extremely well organized and follows an internally consistent style guideline and was very pleasant to review. The package has a clear purpose and achieves it. Most of my comments and suggestions detailed below can be considered generally minor.

- Package passed all checks with no errors, warnings, or notes

- Package passed all tests with no fails, warnings, or skips (964 passes)

- `goodpractice` and `covr` indicate test coverage of 2.57%, though I believe this is incorrect based on my own inspection. The test suite seems to be very robust and complete.

- I would consider adding a few sentences to the README to really explain that the primary set of functions in `stantargets` are target factories, and these functions use the information provided to construct a whole set of targets based on standard workflows/best practices for working with `cmdstanr` and `targets`, by abstracting away much of the technical details of both. The reason `stantargets` makes things easy is because it saves you from having to create and manage many `tar_target()`s manually. I think maybe this point isn't explicitly clear in the README and is important for developing a good mental model of this package. Maybe a simple visualization would be helpful, showing how a single call to a function like `tar_stan_mcmc()` not only creates many targets, but it creates the entire dependency structure (as shown via `tar_visnetwork()`).

- Installing `cmdstan` might be a point of friction for some (though it wasn't for me). You may consider adding some additional text and a link to the README section on `Installation` for where users can find help if they are having installation issues for `cmdstan`/`cmdstanr` since I think it requires a toolchain that users may not already have setup (Example: https://mc-stan.org/cmdstanr/articles/cmdstanr.html)

- Related to the above comment, it might be helpful to very briefly explain the dependency tree that `stantargets` rests on. This may also prevent some confusions for new users about software names and where to find information if installation or other issues arise, since there is `stan`, `cmdstan`, `cmdstanr`, and then `stantargets` to wrap it all together for the purposes of using `targets`. Maybe this will be clearer to others, but I found this a little confusing, at least at first.

- This is super picky, but, should it be made clear in the `Usage` that this is a simplified/toy example to understand the general `stantargets` workflow and not the expected usage. I don't expect many users would want to generate Stan data or define complex functions directly in their `_targets.R` script.

- It would be great if the `Usage` section could perhaps point towards some real
examples, if they exist, similar to `wlandau/targets-keras` _et al._.

- It might be useful to show the full argument names in the `Usage` section for `tar_stan_mcmc()` since this is likely the entry-point to seeing this function used for the first time, and users are unlikely to know the argument order used for positional matching. For example:

  ```r
  list(
    tar_stan_mcmc(
      name = example,
      stan_files = "x.stan",
      data = generate_data()
    )
  )
  ```

- There is some duplication, and syntactic error, in the below text: `tar_read(tar_read(example_summary_x)`

    > Run tar_visnetwork() to check _targets.R for correctness, then call tar_make() to run the pipeline. Access the results using tar_read(), e.g. tar_read(tar_read(example_summary_x). Visit this vignette to read more about this example.

- Consider suppressing the warnings produced by `citation("stantargets")` to clean up this part of the README:

    ````
    ```{r warnings=FALSE}
    citation("stantargets")
    ```
    ````

- Very minor comment: In the following part of the README.md file, you may consider time-stamping the YouTube video (`[this video](https://youtu.be/Gqn7Xn4d5NI?t=401)`) to direct users precisely to the point in the video where the relevant content starts (i.e. ~6min). The easier the better.

  > 1. Basic familiarity with [`targets`](https://docs.ropensci.org/targets/): watch minutes 6 through 40 of [this video](https://youtu.be/Gqn7Xn4d5NI?t=401), then read [this chapter](https://books.ropensci.org/targets/walkthrough.html) of the [user manual](https://books.ropensci.org/targets/).

#### DESCRIPTION/NAMESPACE

- Do the specialized storage formats (e.g. `fst`, `qs`) need to be direct `Imports` for `stantargets`, or can they be `Suggests` and only installed optionally by those using these storage formats? I believe this is how things are handled in `targets`. I tend not to overthink dependencies, but this seems like the right time for scrutiny.
    + Upon further code inspection, I see that `fst_tbl` and `qs` are used as default formats throughout `stantargets`, so maybe these need to stay as Imports

- You have indicated that you plan to submit `stantargets` to CRAN. Could you please reconcile how you plan to submit this package to CRAN while two of the `Imports`, most notably `stan-dev/cmdstanr`, are currently installed from a remote source?

- Only a single function is imported from each of `purrr` (`map()`) and `fs` (`path_rel()`). Is it critical to Import both of these packages for functions that may have base equivalents that are suitable? This could avoid the requirement to install these two packages (as neither package are listed dependencies for `targets`).
    + Actually it looks like `purrr::map2_dfr()` and is also used but is not formally imported like `purrr::map()`

- Following the above comment, most functions from external packages are imported in `tar_stan_package.R`, except is seems to be missing `purrr::map2_dfr`, `rlang::syms`, `rlang::expr`, and `tarchetypes::tar_combine_raw`. Please review.

- You have listed minimum versions of packages for nearly every dependency. Is this strict versioning necessary? Are these minimum versions based on known breaking changes in previous versions, or based on the authors current set of package versions? The latter is considered bad practice by the rOpenSci guidelines (https://devguide.ropensci.org/building.html#pkgdependencies)

- Please consider documenting the data-generating function `tar_stan_example_data()` using the approach recommended here (https://r-pkgs.org/data.html#documenting-data) for documenting standard package data.

- Consider adding `@family` roxygen tags to "families" of functions so they are linked in the documentation, making it easier for users to navigate between functions, perhaps in the same way as the functions are grouped in the `stantargets` `pkgdown` website.

- I think the `@title`, `@description`, and `@return` sections of some functions could use some modest revisions to improve function clarity. I'll address each in the following comments, using `tar_stan_mcmc()` (shown below), as an example:

  ```
  #' @title One MCMC per model with multiple outputs.
  #' @export
  #' @description Targets to run a Stan model once with MCMC
  #'   and save multiple outputs.
  ```

- Since functions are generally meant to represent actions, I suggest the `@title` should be in "future-tense" based on what action will be performed. Since the functions usually just create targets that themselves perform the actions, I think that can be ignored and the function titles can simply describe the actions achieved by the set of child targets, with short and evocative titles being preferred (some current titles may be well-suited as part of the description). The title should not contain any closing punctuation if they are not sentences.

  ```
  #' @title One MCMC per model with multiple outputs.
  ```

  Might be revised to:

  ```
  #' @title Fit one MCMC per model with multiple outputs
  ```

  Or perhaps something similar...

- Often, the `@description` for many functions is ostensibly just listing the return value of that function (usually the function is creating a set of targets), as a sentence fragment. I think it would be preferable to use the `@description` section to expound on what the function does using complete sentences/paragraph styling. Using `tar_stan_mcmc()`, for example, a naive attempt at a revision might look like:

  ```
  #' @description `tar_stan_mcmc()` creates a set of target objects that
  #'  fit a single MCMC for each of the provided model files and
  #'  saves multiple outputs containing results of the model fits.
  ```

- Due to the nature of the `stantargets` functions (i.e. the functions abstract away technical details and behind-the-scenes create and coordinate multiple targets), I think the `@return` section is very important for users to have a solid mental model of what happens when these functions are invoked in a pipeline. I might suggest first describing the return value in a general sense before providing an example of a return, as is currently presented. Expanded on in the next two points.

- For `stantargets` functions that return lists of targets, I like the bullet point list for each of the offspring targets, but the _"descriptions"_ of the targets are a mix of what the target is used for (`reproducibly track the Stan model file.`) and what the target will produce (`data for the generated quantities.`). Consider harmonizing the descriptions or describing each, in turn. See below.

- Following from the above two comments, a revised return section for `tar_stan_mcmc()` _et al._ might look something like this (I've used a markdown table for organization):

  ```
  #' @return
  #' tar_stan_mcmc() returns a list of target objects. The names of the
  #'   targets are prefixed by the `name` argument and suffixed by the
  #'   `stan_files` argument to create unique target names. A set of
  #'   targets is produced for each model file provided to `stan_files`,
  #'   except for the data target, which is shared between models.
  #'
  #' The table below provides an example of the names of the targets
  #'   produced by `tar_stan_gq(name = x, stan_files = "y.stan", ...)`
  #'   with a brief description of each target and what objects (if any)
  #'   are produced when these targets are built.
  #'
  #' | **Name** | **Description** | **Object** |
  #' | :--- | :--- | :--- |
  #' | x_file_y | Track the Stan model file | None |
  #' | x_lines_y | Track contents of the Stan model file (omitted if compile = "original") | file |
  #' | x_data |  Data for the MCMC | data.frame |
  ...and so on
  ```

- Function arguments are defined in an inconsistent format within function documentation. Most commonly, they are defined as `(type) Description`, but also sometimes as `Type, description` and `Type. Description`. With so many arguments for most of the functions, I think it would be best to keep a strict and consistent format, for clarity purposes.

- Again related to function arguments, and I'm not a `cmdstan` expert, but other than the first few required arguments, and the last few `targets`-specific arguments, are the "middle" arguments organized in a thematic way? If not, could they be, or even potentially sorted alphabetically, to make it easier to locate an argument? This is a very minor comment and may be based on my lack of familiarity with all of the moving parts of `Stan` and `cmdstanr`

#### General

- There is some name repetition that I observed when running several of the examples. Very minor but could perhaps be updated to avoid any confusions. It happens because `tempfile()`, by default, names the initial part of the file "file". Using the example in `tar_stan_mcmc()` to demonstrate:

  ```
  ● start target your_model_file_file1b284ba3414
  ● built target your_model_file_file1b284ba3414
  ● start target your_model_data
  ● built target your_model_data
  ● start target your_model_mcmc_file1b284ba3414
  ● built target your_model_mcmc_file1b284ba3414
  ● start target your_model_draws_file1b284ba3414
  ● built target your_model_draws_file1b284ba3414
  ● start target your_model_summary_file1b284ba3414
  ● built target your_model_summary_file1b284ba3414
  ● start target your_model_diagnostics_file1b284ba3414
  ● built target your_model_diagnostics_file1b284ba3414
  ● end pipeline
  ```

  Using something like `tempfile(pattern = "", fileext = ".stan")` makes things cleaner:

  ```
  ● start target your_model_file_X192867a6863e
  ● built target your_model_file_X192867a6863e
  ● start target your_model_data
  ● built target your_model_data
  ● start target your_model_mcmc_X192867a6863e
  ● built target your_model_mcmc_X192867a6863e
  ● start target your_model_diagnostics_X192867a6863e
  ● built target your_model_diagnostics_X192867a6863e
  ● start target your_model_summary_X192867a6863e
  ● built target your_model_summary_X192867a6863e
  ● start target your_model_draws_X192867a6863e
  ● built target your_model_draws_X192867a6863e
  ● end pipeline
  ```

- The example code in `tar_stan_vb_rep_summary()` (shown below) fails locally for me, producing the following error:
  ```r
  targets::tar_dir({ # tar_dir() runs code from a temporary directory.
  targets::tar_script({
  library(stantargets)
  # Do not use temporary storage for stan files in real projects
  # or else your targets will always rerun.
  path <- tempfile(fileext = ".stan")
  tar_stan_example_file(path = path)
  list(
    tar_stan_vb_rep_summary(
      your_model,
      stan_files = path,
      data = tar_stan_example_data(),
      batches = 2,
      reps = 2,
      stdout = R.utils::nullfile(),
      stderr = R.utils::nullfile
    )
  )
  }, ask = FALSE)
  targets::tar_make()
  })
  ```

  ```
  ● start target your_model_batch
  ● built target your_model_batch
  ● start target your_model_file_file5cfc269d8f06
  x error target your_model_file_file5cfc269d8f06
  ● end pipeline
  Error : 'file' must be NULL or an already open connection
  Error: callr subprocess failed: 'file' must be NULL or an already open connection
  Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
  ```

- Should functions (e.g. `tar_stan_mcmc()`, etc.) fail if the file(s) passed to `stan_files` do not exist? This would flag issues when calling functions like `tar_validate()` to assess "correctness"", which seems like the right time to flag this issue, rather than at the time of trying to build the pipeline. But I am not a Stan expert, so maybe there is a reason to not jump the gun on throwing an error when a file doesn't exist. This is what happens if you `tar_make()` with a file that doesn't exist.

  ```
  ● start target name_data
  ● built target name_data
  ● start target name_file_moo
  ● end pipeline
  Error : missing files: model.stan
  Error: callr subprocess failed: missing files: model.stan
  Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
  ```

- `tar_make()` and `tar_make_future()` worked for me without issue in my interactive exploration of the package functionality, but `tar_make_clustermq()` (both the CRAN and GitHub versions) failed, throwing the following errors:

  ```
  Error in try(self$crew$finalize()) : attempt to apply non-function
  ● end pipeline
  Error in self$crew$receive_data() :
    Trying to receive data without workers
  Error: callr subprocess failed: Trying to receive data without workers
  Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
  ```

- Super pedantic: I note that the `trn()` function is used in a lot of places, but at first glance it wasn't clear to me its role and it may be confusing for others trying to understand the code. I very lightly suggest considering a more evocative function name. Upon further inspection, I think it stands for "ternary", as in the ternary operator (similar to the one found in JavaScript and other languages). I believe the term ternary operator is a bit of a misnomer, as it only describes that an operator depends on three operands, but does not describe its actual purpose (conditional flow). Also, in this case, the function is not really an operator, just a standard `R` function.

- For completeness, I had internal dialog about the message queue that I ultimately decided was the concern of `targets`, so I posted a discussion in that repository, found here: https://github.com/ropensci/targets/discussions/391
