### 2021-04-02 #430 @sakrejda
- [X] As the reviewer I confirm that there are no [conflicts of interest](https://devguide.ropensci.org/policies.html#coi) for me to review this work

- [X] **Examples** (that run successfully locally) for all exported functions

---

I don't have any comments beyond those above, I'm looking forward to finding out how well the package does with models that have very large numbers of parameters but didn't have time to check that out for the review.
