### 2021-04-03 e27b755
Briefly explain mental model in the README
