### 2021-04-03 712dcf2
Add help section to README
