+++ README.Rmd
@@ -56,7 +56,7 @@ If you have problems installing [CmdStan](https://github.com/stan-dev/cmdstan), 
 
 ## Usage
 
-First, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate [Stan](https://mc-stan.org/) data, and lists a pipeline of targets. The target list can call target factories like [`tar_stan_mcmc()`](https://wlandau.github.io/stantargets/reference/tar_stan_mcmc.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html).
+First, write a [`_targets.R` file](https://books.ropensci.org/targets/walkthrough.html) that loads your packages, defines a function to generate [Stan](https://mc-stan.org/) data, and lists a pipeline of targets. The target list can call target factories like [`tar_stan_mcmc()`](https://wlandau.github.io/stantargets/reference/tar_stan_mcmc.html) as well as ordinary targets with [`tar_target()`](https://docs.ropensci.org/targets/reference/tar_target.html). The following minimal example is simple enough to contain entirely within the `_targets.R` file, but for larger projects, you may wish to store functions in separate files as in the [`targets-stan`](https://github.com/wlandau/targets-stan) example.
 
 ```{r, eval = FALSE}
 # _targets.R
@@ -71,7 +71,11 @@ generate_data <- function() {
 }
 
 list(
-  tar_stan_mcmc(example, "x.stan", generate_data())
+  tar_stan_mcmc(
+    name = example,
+    stan_files = "x.stan",
+    data = generate_data()
+  )
 )
 ```
 
