+++ README.md
@@ -44,10 +44,11 @@ pipelines](https://wlandau.github.io/stantargets/articles/mcmc.html).
 1.  The [prerequisites of the `targets` R
     package](https://docs.ropensci.org/targets/index.html#prerequisites).
 2.  Basic familiarity with
-    [`targets`](https://docs.ropensci.org/targets/): watch minutes 6
-    through 40 of [this video](https://youtu.be/Gqn7Xn4d5NI), then read
-    [this chapter](https://books.ropensci.org/targets/walkthrough.html)
-    of the [user manual](https://books.ropensci.org/targets/).
+    [`targets`](https://docs.ropensci.org/targets/): watch minutes 7
+    through 40 of [this video](https://youtu.be/Gqn7Xn4d5NI?t=439), then
+    read [this
+    chapter](https://books.ropensci.org/targets/walkthrough.html) of the
+    [user manual](https://books.ropensci.org/targets/).
 3.  Familiarity with Bayesian Statistics and
     [Stan](https://mc-stan.org/). Prior knowledge of
     [`cmdstanr`](https://mc-stan.org/cmdstanr/) helps.
