+++ R/tar_stan_mcmc.R
@@ -38,7 +38,7 @@
 #' @param stan_files Character vector of Stan model files. If you
 #'   supply multiple files, each model will run on the one shared dataset
 #'   generated by the code in `data`. If you supply an unnamed vector,
-#'   `tools::file_path_sans_ext(basename(stan_files))` will be used
+#'   `fs::path_ext_remove(basename(stan_files))` will be used
 #'   as target name suffixes. If `stan_files` is a named vector,
 #'   the suffixed will come from `names(stan_files)`.
 #' @param draws Logical, whether to create a target for posterior draws.
