+++ man/tar_stan_gq.Rd
@@ -50,7 +50,7 @@ Serves as a prefix for target names.}
 \item{stan_files}{Character vector of Stan model files. If you
 supply multiple files, each model will run on the one shared dataset
 generated by the code in \code{data}. If you supply an unnamed vector,
-\code{tools::file_path_sans_ext(basename(stan_files))} will be used
+\code{fs::path_ext_remove(basename(stan_files))} will be used
 as target name suffixes. If \code{stan_files} is a named vector,
 the suffixed will come from \code{names(stan_files)}.}
 
