+++ NAMESPACE
@@ -30,17 +30,20 @@ export(tar_stan_vb_rep_summary)
 export(tar_stan_vb_run)
 importFrom(cmdstanr,cmdstan_model)
 importFrom(digest,digest)
+importFrom(fs,path_ext_remove)
 importFrom(fs,path_rel)
 importFrom(fst,read_fst)
 importFrom(posterior,as_draws_df)
 importFrom(purrr,map)
+importFrom(purrr,map2_dfr)
+importFrom(purrr,map_dbl)
 importFrom(qs,qread)
 importFrom(rlang,check_installed)
 importFrom(rlang,expr)
 importFrom(rlang,quo_squash)
 importFrom(stats,rnorm)
 importFrom(stats,runif)
-importFrom(tarchetypes,tar_combine)
+importFrom(tarchetypes,tar_combine_raw)
 importFrom(tarchetypes,tar_map)
 importFrom(targets,tar_cue)
 importFrom(targets,tar_dir)
