+++ R/tar_stan_example_data.R
@@ -4,6 +4,18 @@
 #'   from [tar_stan_example_file()].
 #' @return List, dataset compatible with the model file from
 #'   [tar_stan_example_file()].
+#' @format A list with the following elements:
+#'   * `n`: integer, number of data points.
+#'   * `x`: numeric, covariate vector.
+#'   * `y`: numeric, response variable.
+#'   * `true_beta`: numeric of length 1, value of the regression
+#'     coefficient `beta` used in simulation.
+#'   * `.join_data`: a list of simulated values to be appended
+#'     to as a `.join_data` column in the output of
+#'     targets generated by functions such as
+#'     [tar_stan_mcmc_rep_summary()]. Contains the
+#'     regression coefficient `beta` (numeric of length 1)
+#'     and prior predictive data `y` (numeric vector).
 #' @param n Integer of length 1, number of data points.
 #' @examples
 #' tar_stan_example_data()
