### 2021-04-04 b746ecf
Sync dependency versions with those of targets
