+++ DESCRIPTION
@@ -34,23 +34,23 @@ Depends:
   R (>= 3.5.0)
 Imports:
   cmdstanr (>= 0.2.0),
-  digest (>= 0.6.21),
+  digest (>= 0.6.25),
   fs (>= 1.5.0),
-  fst (>= 0.9.4),
+  fst (>= 0.9.2),
   posterior (>= 0.1.2),
   purrr (>= 0.3.4),
-  qs (>= 0.14.1),
-  rlang (>= 0.4.8),
+  qs (>= 0.23.2),
+  rlang (>= 0.4.10),
   stats,
   targets (>= 0.0.1),
   tarchetypes (>= 0.0.1),
-  tibble (>= 3.0.4),
+  tibble (>= 3.0.1),
   withr (>= 2.1.2)
 Suggests:
   dplyr (>= 1.0.2),
-  knitr (>= 1.28),
+  knitr (>= 1.30),
   R.utils (>= 2.10.1),
-  rmarkdown (>= 2.1),
+  rmarkdown (>= 2.3),
   testthat (>= 3.0.0),
   visNetwork (>= 2.0.9)
 Remotes:
