### 2021-04-04 6f31948
Error early if Stan model file does not exist
