+++ tests/testthat/test-tar_stan_vb.R
@@ -257,3 +257,22 @@ targets::tar_test("tar_stan_vb(compile = \"copy\") with custom summaries", {
   )
   expect_equal(sort(out), sort(exp))
 })
+
+targets::tar_test("stan files missing", {
+  expect_error(
+    tar_stan_vb(
+      model,
+      stan_files = c("a.stan", "b.stan"),
+      data = c(tar_stan_example_data()),
+      compile = "copy",
+      quiet = TRUE,
+      refresh = 0,
+      iter = 1000,
+      variables = "beta",
+      summaries = list(~quantile(.x, probs = c(0.25, 0.75))),
+      stdout = R.utils::nullfile(),
+      stderr = R.utils::nullfile()
+    ),
+    class = "tar_condition_validate"
+  )
+})
