+++ tests/testthat/test-tar_stan_vb_rep_draws.R
@@ -246,3 +246,25 @@ targets::tar_test("tar_stan_vb_rep_draws(compile = \"copy\") custom", {
   )
   expect_equal(sort(out), sort(exp))
 })
+
+targets::tar_test("stan files missing", {
+  expect_error(
+    tar_stan_vb_rep_draws(
+      model,
+      stan_files = c("a.stan", "b.stan"),
+      data = c(tar_stan_example_data()),
+      compile = "copy",
+      quiet = TRUE,
+      refresh = 0,
+      output_samples = 100,
+      batches = 2,
+      reps = 2,
+      combine = TRUE,
+      data_copy = c("n", "true_beta"),
+      variables = "beta",
+      stdout = R.utils::nullfile(),
+      stderr = R.utils::nullfile()
+    ),
+    class = "tar_condition_validate"
+  )
+})
