+++ R/tar_stan_mle_rep.R
@@ -82,6 +82,7 @@ tar_stan_mle_rep <- function(
   compile <- match.arg(compile)
   assert_chr(stan_files)
   assert_unique(stan_files)
+  lapply(stan_files, assert_stan_file)
   name_stan <- produce_stan_names(stan_files)
   name_file <- paste0(name, "_file")
   name_lines <- paste0(name, "_lines")
