+++ R/tar_stan_vb_rep_summary.R
@@ -40,7 +40,7 @@
 #'     batches = 2,
 #'     reps = 2,
 #'     stdout = R.utils::nullfile(),
-#'     stderr = R.utils::nullfile
+#'     stderr = R.utils::nullfile()
 #'   )
 #' )
 #' }, ask = FALSE)
