### 2021-04-04 1d1d460
Set empty pattern in tempfile()
