+++ R/tar_stan_mcmc_rep.R
@@ -373,7 +373,7 @@ tar_stan_mcmc_rep_run <- function(
   }
   file <- stan_file
   if (identical(compile, "copy")) {
-    tmp <- tempfile(fileext = ".stan")
+    tmp <- tempfile(pattern = "", fileext = ".stan")
     writeLines(stan_file, tmp)
     file <- tmp
   }
