+++ R/tar_stan_gq_rep_summary.R
@@ -1,9 +1,10 @@
 #' @title Multiple runs of generated quantities per model with summaries
 #' @export
-#' @description Targets to run generated quantities multiple times and
+#' @description `tar_stan_gq_rep_summaries()` creates targets
+#'   to run generated quantities multiple times and
 #'   save only the summaries from each run.
 #' @details Most of the arguments are passed to the `$compile()`
-#'   and `$sample()` methods of the `CmdStanModel` class. If you
+#'   and `$generate_quantities()` methods of the `CmdStanModel` class. If you
 #'   previously compiled the model in an upstream [tar_stan_compile()]
 #'   target, then the model should not recompile.
 #' @family generated quantities
