+++ R/tar_stan_gq_rep_draws.R
@@ -1,6 +1,7 @@
 #' @title Multiple runs of generated quantities per model with draws
 #' @export
-#' @description Targets to run generated quantities multiple times and
+#' @description `tar_stan_gq_rep_draws()` creates targets
+#'   to run generated quantities multiple times and
 #'   save only the draws from each run.
 #' @details Most of the arguments are passed to the `$compile()`
 #'   and `$sample()` methods of the `CmdStanModel` class. If you
