+++ R/tar_stan_mcmc.R
@@ -7,8 +7,10 @@
 #'   previously compiled the model in an upstream [tar_stan_compile()]
 #'   target, then the model should not recompile.
 #' @family MCMC
-#' @return `tar_stan_mcmc(name = x, stan_files = "y.stan", ...)` returns a list
-#'   of target objects:
+#' @return A list of target objects. See the "Target objects" section for
+#'   background. The specific target objects returned by
+#'   `tar_stan_mcmc(name = x, stan_files = "y.stan", ...)`
+#'   are as follows.
 #'   * `x_file_y`: reproducibly track the Stan model file.
 #'   * `x_lines_y`: contents of the Stan model file.
 #'     Omitted if `compile = "original"`.
@@ -19,14 +21,7 @@
 #'     Omitted if `summary = FALSE`.
 #'   * `x_diagnostics`: tidy data frame of MCMC sampler diagnostics.
 #'     Omitted if `diagnostics = FALSE`.
-#'  If you supply multiple models, you will get more (model-specific) targets.
-#'  All the models share the same dataset.
-#'
-#'   Target objects represent skippable steps of the analysis pipeline
-#'   as described at <https://books.ropensci.org/targets/>.
-#'   Please see the design specification at
-#'   <https://books.ropensci.org/targets-design/>
-#'   to learn about the structure and composition of target objects.
+#' @inheritSection tar_stan_compile Target objects
 #' @inheritParams cmdstanr::cmdstan_model
 #' @inheritParams cmdstanr::`fit-method-draws`
 #' @inheritParams tar_stan_compile_run
