+++ R/tar_stan_gq_rep_draws.R
@@ -8,8 +8,10 @@
 #'   previously compiled the model in an upstream [tar_stan_compile()]
 #'   target, then the model should not recompile.
 #' @family generated quantities
-#' @return `tar_stan_gq_rep_draws(name = x, stan_files = "y.stan")`
-#'   returns a list of target objects:
+#' @return A list of target objects. See the "Target objects" section for
+#'   background. The specific target objects returned by
+#' `tar_stan_gq_rep_draws(name = x, stan_files = "y.stan")`
+#'   are as follows.
 #'   * `x_file_y`: reproducibly track the Stan model file.
 #'   * `x_lines_y`: contents of the Stan model file.
 #'     Omitted if `compile = "original"`.
@@ -18,12 +20,7 @@
 #'   * `x`: combine all the model-specific draws targets into
 #'     a single data frame with columns to distinguish among the models.
 #'     Suppressed if `combine` is `FALSE`.
-#'
-#'   Target objects represent skippable steps of the analysis pipeline
-#'   as described at <https://books.ropensci.org/targets/>.
-#'   Please see the design specification at
-#'   <https://books.ropensci.org/targets-design/>
-#'   to learn about the structure and composition of target objects.
+#' @inheritSection tar_stan_compile Target objects
 #' @inheritParams tar_stan_gq_rep
 #' @examples
 #' if (Sys.getenv("TAR_LONG_EXAMPLES") == "true") {
