+++ R/tar_stan_mcmc_rep.R
@@ -2,20 +2,10 @@
 #' @keywords internal
 #' @description Internal function for replicated MCMC.
 #'   Users should not invoke directly.
-#' @return `tar_stan_mcmc_rep(name = x, stan_files = "y.stan")`
-#'   returns a list of target objects:
-#'   * `x_file_y`: reproducibly track the Stan model file.
-#'   * `x_lines_y`: contents of the Stan model file.
-#'     Omitted if `compile = "original"`.
-#'   * `x_data`: dynamic branching target with simulated datasets.
-#'   * `x_y`: dynamic branching target with tidy data frames of MCMC summaries.
-#'   * `x`: combine all the model-specific summary targets into
-#'     a single data frame with columns to distinguish among the models.
-#'     Suppressed if `combine` is `FALSE`.
-#'
+#' @return A list of target objects.
 #'   Target objects represent skippable steps of the analysis pipeline
 #'   as described at <https://books.ropensci.org/targets/>.
-#'   Please see the design specification at
+#'   Developers can consult the design specification at
 #'   <https://books.ropensci.org/targets-design/>
 #'   to learn about the structure and composition of target objects.
 #' @inheritParams tar_stan_mcmc_rep_run
