+++ R/tar_stan_gq_rep.R
@@ -2,20 +2,10 @@
 #'   tidy output
 #' @keywords internal
 #' @description Not a user-side function. Do not invoke directly.
-#' @return `tar_stan_gq_rep(name = x, stan_files = "y.stan")`
-#'   returns a list of target objects:
-#'   * `x_file_y`: reproducibly track the Stan model file.
-#'   * `x_lines_y`: contents of the Stan model file.
-#'     Omitted if `compile = "original"`.
-#'   * `x_data`: dynamic branching target with simulated datasets.
-#'   * `x_y`: dynamic branching target with tidy data frames of output.
-#'   * `x`: combine all the model-specific summaries targets into
-#'     a single data frame with columns to distinguish among the models.
-#'     Suppressed if `combine` is `FALSE`.
-#'
+#' @return A list of target objects.
 #'   Target objects represent skippable steps of the analysis pipeline
 #'   as described at <https://books.ropensci.org/targets/>.
-#'   Please see the design specification at
+#'   Developers can consult the design specification at
 #'   <https://books.ropensci.org/targets-design/>
 #'   to learn about the structure and composition of target objects.
 #' @inheritParams tar_stan_gq_rep_run

[…]
