+++ R/tar_stan_vb_rep_draws.R
@@ -12,7 +12,8 @@
 #'   previously compiled the model in an upstream [tar_stan_compile()]
 #'   target, then the model should not recompile.
 #' @family variational Bayes
-#' @return A list of target objects. See the "Target objects" section for
+#' @return `tar_stan_vb_rep_summaries()` returns a
+#'   list of target objects. See the "Target objects" section for
 #'   background. The specific target objects returned by
 #'   `tar_stan_vb_rep_draws(name = x, stan_files = "y.stan")`
 #'   are as follows.
