+++ R/tar_stan_gq_rep_summary.R
@@ -13,7 +13,7 @@
 #'   background.
 #'   The target names use the `name` argument as a prefix, and the individual
 #'   elements of `stan_files` appear in the suffixes where applicable.
-#'   The specific target objects returned by
+#'   As an example, the specific target objects returned by
 #'   `tar_stan_gq_rep_summary(name = x, stan_files = "y.stan")`
 #'   returns a list of target objects:
 #'   * `x_file_y`: reproducibly track the Stan model file.
