+++ R/tar_stan_gq.R
@@ -17,13 +17,20 @@
 #'   As an example, the specific target objects returned by
 #'   `tar_stan_gq(name = x, stan_files = "y.stan", ...)`
 #'   are as follows.
-#'   * `x_file_y`: reproducibly track the Stan model file.
-#'   * `x_lines_y`: contents of the Stan model file.
-#'     Omitted if `compile = "original"`.
-#'   * `x_data`: data for the generated quantities.
-#'   * `x_gq_y`: `CmdStanGQ` object with all the generated quantities results.
-#'   * `x_draws_y`: tidy data frame of draws. Omitted if `draws = FALSE`.
-#'   * `x_summary_y`: tidy data frame of summaries.
+#'   * `x_file_y`: reproducibly track the Stan model file. Returns
+#'     a character vector with the model file and compiled executable.
+#'   * `x_lines_y`: read the Stan model file for safe transport to
+#'     parallel workers. Omitted if `compile = "original"`.
+#'     Returns a character vector of lines in the model file.
+#'   * `x_data`: run the R expression in the `data` argument to produce
+#'     a Stan dataset for the model. Returns a Stan data list.
+#'   * `x_gq_y`: run generated quantities on the model and the dataset.
+#'     Returns a `cmdstanr` `CmdStanGQ` object with all the results.
+#'   * `x_draws_y`: extract draws from `x_gq_y`.
+#'     Omitted if `draws = FALSE`.
+#'     Returns a tidy data frame of draws.
+#'   * `x_summary_y`: extract compact summaries from `x_gq_y`.
+#'     Returns a tidy data frame of summaries.
 #'     Omitted if `summary = FALSE`.
 #' @inheritSection tar_stan_compile Target objects
 #' @inheritParams cmdstanr::cmdstan_model
