### 2021-04-04 #430 @wlandau
## Respnose to @mattwarkentin's review

The documentation of `stantargets` is so much better now, thanks again @mattwarkentin for your feedback.

> goodpractice and covr indicate test coverage of 2.57%, though I believe this is incorrect based on my own inspection.

That's odd. `stantargets` does skip tests if:

* `dplyr` is not installed, or
* CmdStan is not installed, or
* The developer sets the `TAR_STAN_SKIP_TEST_COMPILE_COPY` environment variable to `"true"`. (Compilation is a testing bottleneck, so sometimes I skip the tests that require it.)

Could that explain it?

> I would consider adding a few sentences to the README to really explain that the primary set of functions in stantargets are target factories, and these functions use the information provided to construct a whole set of targets based on standard workflows/best practices for working with cmdstanr and targets, by abstracting away much of the technical details of both. The reason stantargets makes things easy is because it saves you from having to create and manage many tar_target()s manually. I think maybe this point isn't explicitly clear in the README and is important for developing a good mental model of this package. Maybe a simple visualization would be helpful, showing how a single call to a function like tar_stan_mcmc() not only creates many targets, but it creates the entire dependency structure (as shown via tar_visnetwork()).

Addressed in https://github.com/wlandau/stantargets/commit/e27b7552ce8dd3282950bd01090537136cf94ca4. The README now gives a brief explanation and refers readers to a vignette for an example.

> Installing cmdstan might be a point of friction for some (though it wasn't for me). You may consider adding some additional text and a link to the README section on Installation for where users can find help if they are having installation issues for cmdstan/cmdstanr since I think it requires a toolchain that users may not already have setup (Example: https://mc-stan.org/cmdstanr/articles/cmdstanr.html)

Addressed in https://github.com/wlandau/stantargets/commit/016dd5bb4b5fc1fee9d7dfd686f6ddee97674b8e.

> Related to the above comment, it might be helpful to very briefly explain the dependency tree that stantargets rests on. This may also prevent some confusions for new users about software names and where to find information if installation or other issues arise, since there is stan, cmdstan, cmdstanr, and then stantargets to wrap it all together for the purposes of using targets. Maybe this will be clearer to others, but I found this a little confusing, at least at first.

https://github.com/wlandau/stantargets/commit/712dcf274b45a87c005a10c46735f3eb70c6de5d mentions this in the context of troubleshooting, which is where I have seen  confusion come up most often.

> This is super picky, but, should it be made clear in the Usage that this is a simplified/toy example to understand the general stantargets workflow and not the expected usage. I don't expect many users would want to generate Stan data or define complex functions directly in their _targets.R script.

Clarified in https://github.com/wlandau/stantargets/commit/0df9acafbe720bc702a260913c9f8b09cf31e372.

> It would be great if the Usage section could perhaps point towards some real
examples, if they exist, similar to wlandau/targets-keras et al..

Good idea. I adapted https://github.com/wlandau/targets-stan and created a project that will eventually appear at https://github.com/wlandau/stantargets-example-validation. Unfortunately, it will not become public for another week or two because I need permission from my company to release it and the gatekeepers are on vacation.

> It might be useful to show the full argument names in the Usage section for tar_stan_mcmc() since this is likely the entry-point to seeing this function used for the first time, and users are unlikely to know the argument order used for positional matching. For example:
> list(
>  tar_stan_mcmc(
>    name = example,
>    stan_files = "x.stan",
>    data = generate_data()
>  )
> )

Done in https://github.com/wlandau/stantargets/commit/0c14616c01aa642292f021db986a50f87abde860.

> There is some duplication, and syntactic error, in the below text: tar_read(tar_read(example_summary_x)
> Run tar_visnetwork() to check _targets.R for correctness, then call tar_make() to run the pipeline. Access the results using tar_read(), e.g. tar_read(tar_read(example_summary_x). Visit this vignette to read more about this example.

Fixed in https://github.com/wlandau/stantargets/commit/f4265305ac05df53543438dc0c5c9883dabe9115.

> Consider suppressing the warnings produced by citation("stantargets") to clean up this part of the README:
> ```{r warnings=FALSE}
> citation("stantargets")
> ```

Fixed in https://github.com/wlandau/stantargets/commit/708379d7e47be609780024960519120464b36e98.

> Very minor comment: In the following part of the README.md file, you may consider time-stamping the YouTube video ([this video](https://youtu.be/Gqn7Xn4d5NI?t=401)) to direct users precisely to the point in the video where the relevant content starts (i.e. ~6min). The easier the better.

Added in https://github.com/wlandau/stantargets/commit/aa3ecbfd7e8d1d13746a15fa52303e08ccb44cde.

> Do the specialized storage formats (e.g. fst, qs) need to be direct Imports for stantargets, or can they be Suggests and only installed optionally by those using these storage formats? I believe this is how things are handled in targets. I tend not to overthink dependencies, but this seems like the right time for scrutiny...Upon further code inspection, I see that fst_tbl and qs are used as default formats throughout stantargets, so maybe these need to stay as Imports.

Yes, `stantargets` functions automatically choose `"fst_tbl"` and `"qs"` so the user does not have to think about storage formats. Since these formats are always used, I think the corresponding packages should be hard dependencies.

> You have indicated that you plan to submit stantargets to CRAN. Could you please reconcile how you plan to submit this package to CRAN while two of the Imports, most notably stan-dev/cmdstanr, are currently installed from a remote source?

The CRAN release of `stantargets` is going to be delayed for as long as it takes `cmdstanr` and `posterior` to reach CRAN. The Stan Development Team has expressed the desire to release both these packages eventually, but I do not know exactly when that will happen. But I do plan to publish `stantargets` as soon as its dependencies are ready.

> Only a single function is imported from each of purrr (map()) and fs (path_rel()). Is it critical to Import both of these packages for functions that may have base equivalents that are suitable? This could avoid the requirement to install these two packages (as neither package are listed dependencies for targets)...Actually it looks like purrr::map2_dfr() and is also used but is not formally imported like purrr::map(). Following the above comment, most functions from external packages are imported in tar_stan_package.R, except is seems to be missing purrr::map2_dfr, rlang::syms, rlang::expr, and tarchetypes::tar_combine_raw. Please review.

I fixed some of this in @sakrejda's review. The rest should be fixed in https://github.com/wlandau/stantargets/commit/c83b1d7748a1873c36090ef3e34a288c6ec82b69. I feel `stantargets` operates at a high enough level where it makes sense to use `fs` and `purrr` sometimes, and that commit imports more functions that should have been imported (including `fs::path_ext_remove()` rather than `tools::file_path_sans_ext()`, which removes the explicit dependency on `tools`).

> You have listed minimum versions of packages for nearly every dependency. Is this strict versioning necessary? Are these minimum versions based on known breaking changes in previous versions, or based on the authors current set of package versions? The latter is considered bad practice by the rOpenSci guidelines (https://devguide.ropensci.org/building.html#pkgdependencies)

That's a new one for me. And I suspect it may reflect a recent change in the R community's overall thinking. I definitely remember looking at `DESCRIPTION` files of major Tidyverse packages a few years ago and noticing that they set minimum versions for all the dependencies.

I see where this guideline is coming from, and I gave it serious thought. However, I am reluctant to drop the minimum version requirements of `targets` or any of its derivatives. Reasons:

1. Years of bug reports for both `drake` and `targets` motivated many of these minimum version requirements. I am certain there were good reasons for `rlang`, `digest`, `qs`, and `tibble`, but I was not proactive about documenting them at the time, so all the current minimum version requirements are [Chesterton's Fences](https://wiki.lesswrong.com/wiki/Chesterton%27s_Fence).
2. As you pointed out, `stantargets` has a lot of dependencies, which creates a lot of room for things to go wrong. `targets` frequently gets blamed for problems in other packages, so I end up spending a lot of time on tech support outside the scope of the packages I actually develop. Minimum version requirements help mitigate this.
3. `stantargets` belongs in [project-oriented workflows](https://rstats.wtf/project-oriented-workflow.html) where users should be using `renv` anyway. That should help avoid modifications to the global package environment.

I did take another look at the minimum version requirements of `stantargets`, and in https://github.com/wlandau/stantargets/commit/b746ecfd695c7d8db14b2e22085a2e8f2bb8608a, I synchronized them with those of `targets`.

> Please consider documenting the data-generating function tar_stan_example_data() using the approach recommended here (https://r-pkgs.org/data.html#documenting-data) for documenting standard package data.

Added a `@format` tag in https://github.com/wlandau/stantargets/commit/1e9f083a40d712199d35df4526d00e2b61f01443.

> Consider adding @family roxygen tags to "families" of functions so they are linked in the documentation, making it easier for users to navigate between functions, perhaps in the same way as the functions are grouped in the stantargets pkgdown website.

Yeah, that's much more convenient than `@seealso`. Implemented in https://github.com/wlandau/stantargets/commit/1d103b07e45aa20528fcb9496683c920a4252975 to reflect the groupings at https://wlandau.github.io/stantargets/reference/index.html.

>I think the @title, @description, and @return sections of some functions could use some modest revisions to improve function clarity. I'll address each in the following comments, using tar_stan_mcmc() (shown below), as an example:
>
>#' @title One MCMC per model with multiple outputs.
#' @export
#' @description Targets to run a Stan model once with MCMC
#'   and save multiple outputs.
Since functions are generally meant to represent actions, I suggest the @title should be in "future-tense" based on what action will be performed. Since the functions usually just create targets that themselves perform the actions, I think that can be ignored and the function titles can simply describe the actions achieved by the set of child targets, with short and evocative titles being preferred (some current titles may be well-suited as part of the description). The title should not contain any closing punctuation if they are not sentences.
>
> #' @title One MCMC per model with multiple outputs.
> Might be revised to:
>
> #' @title Fit one MCMC per model with multiple outputs
> Or perhaps something similar...

I removed terminating punctuation from the `@title`s as you suggested in https://github.com/wlandau/stantargets/commit/1d103b07e45aa20528fcb9496683c920a4252975.

To me, a title like "fit one MCMC per model with multiple outputs" implies that the function actually does the action instead of creating a target to do it later. I worry this may confuse new users of `targets` who often struggle to understand that calling `tar_target(name, command())` does not actually run `command()`.

> Often, the @description for many functions is ostensibly just listing the return value of that function (usually the function is creating a set of targets), as a sentence fragment. I think it would be preferable to use the @description section to expound on what the function does using complete sentences/paragraph styling. Using tar_stan_mcmc(), for example, a naive attempt at a revision might look like:
>
>#' @description `tar_stan_mcmc()` creates a set of target objects that
>#'  fit a single MCMC for each of the provided model files and
>#'  saves multiple outputs containing results of the model fits.

I agree, edited in https://github.com/wlandau/stantargets/commit/a6de8786530c34cbc4e652cc22f786c64dd912fb.

> Due to the nature of the stantargets functions (i.e. the functions abstract away technical details and behind-the-scenes create and coordinate multiple targets), I think the @return section is very important for users to have a solid mental model of what happens when these functions are invoked in a pipeline. I might suggest first describing the return value in a general sense before providing an example of a return, as is currently presented. Expanded on in the next two points.

https://github.com/wlandau/stantargets/commit/ec0d5e561abe85445550d19a76a4be26fe4e007f introduces a new `@section` on target objects. All target factories now inherit this section and refer to it at the beginning of `@return`. (See also https://github.com/wlandau/stantargets/commit/3822b96bb9cd05fa1add471aaf355a9df5b8e9c7).

> For stantargets functions that return lists of targets, I like the bullet point list for each of the offspring targets, but the "descriptions" of the targets are a mix of what the target is used for (reproducibly track the Stan model file.) and what the target will produce (data for the generated quantities.). Consider harmonizing the descriptions or describing each, in turn. See below.
>
> Following from the above two comments, a revised return section for tar_stan_mcmc() et al. might look something like this (I've used a markdown table for organization):
>
#' @return
> #' tar_stan_mcmc() returns a list of target objects. The names of the
#'   targets are prefixed by the `name` argument and suffixed by the
#'   `stan_files` argument to create unique target names. A set of
#'   targets is produced for each model file provided to `stan_files`,
#'   except for the data target, which is shared between models.
#'
#' The table below provides an example of the names of the targets
#'   produced by `tar_stan_gq(name = x, stan_files = "y.stan", ...)`
#'   with a brief description of each target and what objects (if any)
#'   are produced when these targets are built.
#'
#' | **Name** | **Description** | **Object** |
#' | :--- | :--- | :--- |
#' | x_file_y | Track the Stan model file | None |
#' | x_lines_y | Track contents of the Stan model file (omitted if compile = "original") | file |
#' | x_data |  Data for the MCMC | data.frame |
...and so on

https://github.com/wlandau/stantargets/commit/884f36f0113cbcd295d1e54e1b57e4ef9490bcfd describes the returned targets with more consistency and more detail. (See also https://github.com/wlandau/stantargets/commit/c688f1f1e551d1ab685cb86129e5be1797e2cf9f and https://github.com/wlandau/stantargets/commit/65672acaa751b73b50ee4a7524822e09bf35abbd.) I would have liked to use Markdown tables as you suggested, but they did not render properly.

> Function arguments are defined in an inconsistent format within function documentation. Most commonly, they are defined as (type) Description, but also sometimes as Type, description and Type. Description. With so many arguments for most of the functions, I think it would be best to keep a strict and consistent format, for clarity purposes.

Unfortunately, I do not think I can fix this one. The arguments defined with "(type) Description" are inherited directly from the `cmdstanr` documentation (e.g. `@inheritParams cmdstanr::cmdstan_model`). The others are implemented directly in `stantargets` or inherited from `targets`. To rectify the inconsistency, I would have to change all the inherited arguments in `targets`, which will probably create other instances of the same inconsistency in other [R Targetopia](https://wlandau.github.io/targetopia/packages.html) packages.

> Again related to function arguments, and I'm not a cmdstan expert, but other than the first few required arguments, and the last few targets-specific arguments, are the "middle" arguments organized in a thematic way? If not, could they be, or even potentially sorted alphabetically, to make it easier to locate an argument? This is a very minor comment and may be based on my lack of familiarity with all of the moving parts of Stan and cmdstanr

The middle arguments are mostly ordered the same way as in the analogous functions in `cmdstanr` (e.g. [`tar_stan_mcmc()`](https://wlandau.github.io/stantargets/reference/tar_stan_mcmc.html) vs [`model-method-compile`](https://mc-stan.org/cmdstanr/reference/model-method-compile.html) and [`model-method-sample`](https://mc-stan.org/cmdstanr/reference/model-method-sample.html)). I feel this eases the transition from `cmdstanr` to `stantargets`. (But bome `stantargets` arguments are inserted next to their thematically appropriate counterparts, such as `stdout` and `stderr`, which are next to `quiet`.)

Generally speaking, in the signatures of target factories, I order arguments as follows:

1. Required `stantargets`-specific arguments.
2. `cmdstanr` arguments for compilation.
3. `cmdstanr` arguments to run the Bayesian method.
4. Optional `stantargets`-specific arguments.
5. Optional `targets`-specific arguments.

The 5 groups above are in rough order of priority, and (2) goes before (3) because (3) has a massive number of arguments. Within each group, users of `cmdstanr` and `targets` are already accustomed to seeing those arguments next to each other.

> There is some name repetition that I observed when running several of the examples. Very minor but could perhaps be updated to avoid any confusions. It happens because tempfile(), by default, names the initial part of the file "file". Using the example in tar_stan_mcmc() to demonstrate:
>
> ● start target your_model_file_file1b284ba3414
● built target your_model_file_file1b284ba3414
● start target your_model_data
● built target your_model_data
● start target your_model_mcmc_file1b284ba3414
● built target your_model_mcmc_file1b284ba3414
● start target your_model_draws_file1b284ba3414
● built target your_model_draws_file1b284ba3414
● start target your_model_summary_file1b284ba3414
● built target your_model_summary_file1b284ba3414
● start target your_model_diagnostics_file1b284ba3414
● built target your_model_diagnostics_file1b284ba3414
● end pipeline
Using something like tempfile(pattern = "", fileext = ".stan") makes things cleaner:
>
> ● start target your_model_file_X192867a6863e
● built target your_model_file_X192867a6863e
● start target your_model_data
● built target your_model_data
● start target your_model_mcmc_X192867a6863e
● built target your_model_mcmc_X192867a6863e
● start target your_model_diagnostics_X192867a6863e
● built target your_model_diagnostics_X192867a6863e
● start target your_model_summary_X192867a6863e
● built target your_model_summary_X192867a6863e
● start target your_model_draws_X192867a6863e
● built target your_model_draws_X192867a6863e
● end pipeline

Sure, fixed in https://github.com/wlandau/stantargets/commit/1d1d4605d5c49efa62ae96f95b1ccbd0ae1cc1ca.

>The example code in tar_stan_vb_rep_summary() (shown below) fails locally for me, producing the following error:
>
> targets::tar_dir({ # tar_dir() runs code from a temporary directory.
targets::tar_script({
library(stantargets)
# or else your targets will always rerun.
path <- tempfile(fileext = ".stan")
tar_stan_example_file(path = path)
[…]
