+++ DESCRIPTION
@@ -27,6 +27,16 @@ Authors@R: c(
     comment = c(ORCID = "0000-0003-1878-3253")
   ),
   person(
+    given = "Krzysztof",
+    family = "Sakrejda",
+    role = "rev"
+  ),
+  person(
+    given = c("Matthew", "T."),
+    family = "Warkentin",
+    role = "rev"
+  ),
+  person(
     family = "Eli Lilly and Company",
     role = "cph"
   ))
