### 2021-04-05 #430 @sakrejda
> I will keep this in mind and think about a new name. However, I hesitate to change it immediately because I use trn() in most of my other packages, and updating them all would take time.

Maybe documenting it would help even though it's a utility? Here's what I did during my review:

```
> library(stantargets)
> ?trn
No documentation for ‘trn’ in specified packages and libraries:
you could try ‘??trn’
> trn
Error: object 'trn' not found
> stantargets::trn
Error: 'trn' is not an exported object from 'namespace:stantargets'
> stantargets:::trn
function (condition, x, y)
{
    if (any(condition)) {
        x
    }
    else {
        y
    }
}
<bytecode: 0x55d183e9a9a8>
<environment: namespace:stantargets>
>
```

Bit of a brainfart on my part but doc would help :shrug:
