### 2021-04-05 #430 @tjmahr
just observing the that one alternative name for `trn()`, based on the source, is `ifany()`
