### 2021-04-05 bfa2098
trn() => if_any()
