+++ R/tar_stan_mle.R
@@ -146,7 +146,7 @@ tar_stan_mle <- function(
   )
   args_mle <- list(
     call_ns("stantargets", "tar_stan_mle_run"),
-    stan_file = trn(identical(compile, "original"), sym_file, sym_lines),
+    stan_file = if_any(identical(compile, "original"), sym_file, sym_lines),
     data = sym_data,
     compile = compile,
     quiet = quiet,
