+++ R/tar_stan_mcmc.R
@@ -193,7 +193,7 @@ tar_stan_mcmc <- function(
   )
   args_mcmc <- list(
     call_ns("stantargets", "tar_stan_mcmc_run"),
-    stan_file = trn(identical(compile, "original"), sym_file, sym_lines),
+    stan_file = if_any(identical(compile, "original"), sym_file, sym_lines),
     data = sym_data,
     compile = compile,
     quiet = quiet,
