### 2021-04-05 #430 @mattwarkentin
@wlandau Thanks for providing such detailed responses. A few remaining comments:

- There is still a minor spelling error that didn't quite get fixed in https://github.com/wlandau/stantargets/commit/f4265305ac05df53543438dc0c5c9883dabe9115
    + See [here](https://github.com/wlandau/stantargets/blob/bfa20986ca1ba10d972da067cc0dd6c2507d5670/README.Rmd#L82): "e.g. `ar_read(example_summary_x)`"

> If you have time, could we troubleshoot in a discussion thread? It would help if you would peel back some layers in a small reprex. Could just be a problem with targets or clustermq. Personally, I use stantargets with tar_make_clustermq() a lot at work without this error.

Yes, I'm happy to try and troubleshoot this issue.  I will try to put something together tomorrow when I'm back in the office. It _may_ just be a me issue, but we can find that out.

@melvidoni I'm not sure if you need/want me to _"sign off"_ on anything, but @wlandau has more than adequately addressed all of my comments. I am happy for this review to proceed forward to whatever is next.
