+++ README.Rmd
@@ -79,7 +79,7 @@ list(
 )
 ```
 
-Run [`tar_visnetwork()`](https://docs.ropensci.org/targets/reference/tar_visnetwork.html) to check `_targets.R` for correctness, then call [`tar_make()`](https://docs.ropensci.org/targets/reference/tar_make.html) to run the pipeline. Access the results using [`tar_read()`](https://docs.ropensci.org/targets/reference/tar_read.html), e.g. `ar_read(example_summary_x)`. Visit [this vignette](https://wlandau.github.io/stantargets/articles/mcmc.html) to read more about this example.
+Run [`tar_visnetwork()`](https://docs.ropensci.org/targets/reference/tar_visnetwork.html) to check `_targets.R` for correctness, then call [`tar_make()`](https://docs.ropensci.org/targets/reference/tar_make.html) to run the pipeline. Access the results using [`tar_read()`](https://docs.ropensci.org/targets/reference/tar_read.html), e.g. `tar_read(example_summary_x)`. Visit [this vignette](https://wlandau.github.io/stantargets/articles/mcmc.html) to read more about this example.
 
 ## Help
 
