### 2021-04-05 #430 @wlandau
Thanks again for reviewing, @mattwarkentin! The spelling mistake is fixed in https://github.com/wlandau/stantargets/commit/296fcac37713b2e8cb6e8e6ed4379c8910c2c57c.
