### 2021-04-06 #430 @melvidoni
Approved! Thanks @wlandau  for submitting and @mattwarkentin @sakrejda  for your reviews!

To-dos for @wlandau:
    * remove styling tweaks from your `pkgdown` config but keep that config file
    * replace your package docs URL with `https://docs.ropensci.org/package_name`
- [ ] Fix any links in badges for CI and coverage to point to the ropensci URL. We no longer transfer Appveyor projects to ropensci Appveyor account so after transfer of your repo to rOpenSci's "ropensci" GitHub organization the badge should be `[![AppVeyor Build Status](https://ci.appveyor.com/api/projects/status/github/ropensci/pkgname?branch=master&svg=true)](https://ci.appveyor.com/project/individualaccount/pkgname)`. If Appveyor does not pick up new commits after transfer, you might need to delete and re-create the Appveyor project. (Repo transfers are smoother with GitHub Actions)
- [ ] Please check you updated the package version post-review version updated and that you documented all changes in NEWS.md
