### 2021-04-09 63c62e0
Edit links and use rOpenSci code of conduct
