+++ .github/ISSUE_TEMPLATE/performance.md
@@ -9,9 +9,9 @@ assignees: wlandau
 
 ## Prework
 
-* [ ] Read and agree to the [code of conduct](https://github.com/wlandau/stantargets/blob/main/CODE_OF_CONDUCT.md) and [contributing guidelines](https://github.com/wlandau/stantargets/blob/main/CONTRIBUTING.md).
-* [ ] If there is [already a relevant issue](https://github.com/wlandau/stantargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
-* [ ] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/wlandau/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
+* [ ] Read and agree to the [code of conduct](https://ropensci.org/code-of-conduct/) and [contributing guidelines](https://github.com/ropensci/stantargets/blob/main/CONTRIBUTING.md).
+* [ ] If there is [already a relevant issue](https://github.com/ropensci/stantargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
+* [ ] Post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/ropensci/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot the problems you identify. A reproducible example is:
     * [ ] **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
     * [ ] **Minimal**: reduce runtime wherever possible and remove complicated details that are irrelevant to the issue at hand.
     * [ ] **Readable**: format your code according to the [tidyverse style guide](https://style.tidyverse.org/).
