+++ .github/ISSUE_TEMPLATE/feature.md
@@ -8,8 +8,8 @@ assignees: wlandau
 
 ## Prework
 
-* [ ] I understand and agree to the [code of conduct](https://github.com/wlandau/stantargets/blob/main/CODE_OF_CONDUCT.md) and  [contributing guidelines](https://github.com/wlandau/stantargets/blob/main/CONTRIBUTING.md).
-* [ ] If there is [already a relevant issue](https://github.com/wlandau/stantargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
+* [ ] I understand and agree to the [code of conduct](https://ropensci.org/code-of-conduct/) and  [contributing guidelines](https://github.com/ropensci/stantargets/blob/main/CONTRIBUTING.md).
+* [ ] If there is [already a relevant issue](https://github.com/ropensci/stantargets/issues), whether open or closed, comment on the existing thread instead of posting a new issue.
 * [ ] New features take time and effort to create, and they take even more effort to maintain. So if the purpose of the feature is to resolve a struggle you are encountering personally, please consider first posting a "trouble" or "other" issue so we can discuss your use case and search for existing solutions first.
 
 ## Proposal
