+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,9 +1,8 @@
 # Prework
 
-* [ ] I understand and agree to this repository's [code of conduct](https://github.com/wlandau/stantargets/blob/main/CODE_OF_CONDUCT.md).
-* [ ] I understand and agree to this repository's [contributing guidelines](https://github.com/wlandau/stantargets/blob/main/CONTRIBUTING.md).
-* [ ] I have already submitted an issue to the [issue tracker](http://github.com/wlandau/stantargets/issues) to discuss my idea with the maintainer.
-* [ ] This pull request is not a [draft](https://github.blog/2019-02-14-introducing-draft-pull-requests).
+* [ ] I understand and agree to this repository's [code of conduct](https://ropensci.org/code-of-conduct/).
+* [ ] I understand and agree to this repository's [contributing guidelines](https://github.com/ropensci/stantargets/blob/main/CONTRIBUTING.md).
+* [ ] I have already submitted an issue to the [issue tracker](http://github.com/ropensci/stantargets/issues) to discuss my idea with the maintainer.
 
 # Related GitHub issues and pull requests
 
