+++ CONTRIBUTING.md
@@ -4,23 +4,16 @@ Development is a community effort, and we welcome participation.
 
 ## Code of Conduct
 
-By participating in this project, you agree to abide by the [code of conduct](https://github.com/wlandau/stantargets/blob/main/CODE_OF_CONDUCT.md).
+By participating in this project, you agree to abide by the [code of conduct](https://ropensci.org/code-of-conduct/).
 
 ## Issues
 
-Anyone can submit an issue to <https://github.com/wlandau/stantargets/issues> or comment on an existing thread. Common reasons are to
+Anyone can start or contribute to an [issue](https://github.com/ropensci/stantargets/issues) or [discussion thread](https://github.com/ropensci/stantargets/discussions). Issues are mainly for bug reports and package maintenance, and discussions are for usage help and brainstorming. Please respect the following guidelines.
 
-* Suggest features.
-* Report bugs.
-* Request help with difficult use cases.
-* Ask questions.
-
-Please abide by the following guidelines.
-
-* Before posting a new issue, please take a moment to search for existing similar issues in order to avoid duplication.
-* For bug reports: if you can, please install the latest GitHub version of `stantargets` (i.e. `remotes::install_github("wlandau/stantargets")`) and verify that the issue still persists.
+* Before posting a new issue or discussion, please take a moment to search for existing threads in order to avoid duplication.
+* For bug reports: if you can, please install the latest GitHub version of `stantargets` (i.e. `remotes::install_github("ropensci/stantargets")`) and verify that the issue still persists.
 * Describe your issue in prose as clearly and concisely as possible.
-* For any problem you identify, post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/wlandau/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot. A reproducible example is:
+* For any problem you identify, post a [minimal reproducible example](https://www.tidyverse.org/help/) like [this one](https://github.com/ropensci/targets/issues/256#issuecomment-754229683) so the maintainer can troubleshoot. A reproducible example is:
     * **Runnable**: post enough R code and data so any onlooker can create the error on their own computer.
     * **Minimal**: reduce runtime wherever possible and remove complicated details that are irrelevant to the issue at hand.
     * **Readable**: format your code according to the [tidyverse style guide](https://style.tidyverse.org/).
@@ -29,7 +22,7 @@ Please abide by the following guidelines.
 
 External code contributions are extremely helpful in the right circumstances. Here are the recommended steps.
 
-1. Prior to contribution, please propose your idea in a [new issue thread](https://github.com/wlandau/stantargets/issues) so you and the maintainer can define the intent and scope of your work.
+1. Prior to contribution, please propose your idea in a [new issue thread](https://github.com/ropensci/stantargets/issues) so you and the maintainer can define the intent and scope of your work.
 2. [Fork the repository](https://help.github.com/articles/fork-a-repo/).
 3. Follow the [GitHub flow](https://guides.github.com/introduction/flow/index.html) to create a new branch, add commits, and open a pull request.
 4. Discuss your code with the maintainer in the pull request thread.
@@ -42,5 +35,5 @@ Please also follow these additional guidelines.
 * Format your code according to the [tidyverse style guide](https://style.tidyverse.org/) and check your formatting with the `lint_package()` function from the [`lintr`](https://github.com/jimhester/lintr) package.
 * Check code coverage with `covr::package_coverage()`. Automated tests should cover all the new or changed functionality in your pull request.
 * Run overall package checks with `devtools::check()` and `goodpractice::gp()`
[…]
