+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,14 +1,14 @@
 lank_issues_enabled: true
 contact_links:
   - name: General
-    url: https://github.com/wlandau/stantargets/discussions/categories/general
+    url: https://github.com/ropensci/stantargets/discussions/categories/general
     about: Chat about stantargets.
   - name: Help
-    url: https://github.com/wlandau/stantargets/discussions/categories/help
+    url: https://github.com/ropensci/stantargets/discussions/categories/help
     about: Ask questions and get help.
   - name: Ideas
-    url: https://github.com/wlandau/stantargets/discussions/categories/ideas
+    url: https://github.com/ropensci/stantargets/discussions/categories/ideas
     about: Brainstorm ideas for new features or use cases.
   - name: Show and tell
-    url: https://github.com/wlandau/stantargets/discussions/categories/show-and-tell
+    url: https://github.com/ropensci/stantargets/discussions/categories/show-and-tell
     about: Show off something you have made.

[…]
