+++ inst/paper.bib
@@ -63,7 +63,7 @@
   stantargets,
   title = {stantargets: Targets for Stan Workflows},
   author = {William Michael Landau},
-  note = {https://wlandau.github.io/stantargets/, https://github.com/wlandau/stantargets},
+  note = {https://docs.ropensci.org/stantargets/, https://github.com/ropensci/stantargets},
 }
 
 @Manual{
