+++ R/tar_stan_mcmc_rep.R
@@ -23,7 +23,7 @@
 #'   summaries, include a `.join_data`
 #'   element of your Stan data list with names and dimensions corresponding
 #'   to those of the model. For details, read
-#'   <https://wlandau.github.io/stantargets/articles/mcmc_rep.html>.
+#'   <https://docs.ropensci.org/stantargets/articles/mcmc_rep.html>.
 #' @param batches Number of batches. Each batch is a sequence
 #'   of branch targets containing multiple reps. Each rep
 #'   generates a dataset and runs the model on it.
@@ -305,7 +305,7 @@ tar_stan_mcmc_rep <- function(
 #'   for each rep. To join more than just scalars, include a `.join_data`
 #'   element of your Stan data list with names and dimensions corresponding
 #'   to those of the model. For details, read
-#'   <https://wlandau.github.io/stantargets/articles/mcmc_rep.html>.
+#'   <https://docs.ropensci.org/stantargets/articles/mcmc_rep.html>.
 tar_stan_mcmc_rep_run <- function(
   stan_file,
   stan_name,
