+++ NEWS.md
@@ -1,9 +1,23 @@
+# stantargets 0.0.0.9003
+
+* Skip tests if CmdStan is not installed (@sakrejda).
+* Use custom `generate_data()` function in the docs, as opposed to `tar_stan_example_data()` directly (@sakrejda).
+* Add the `pedantic` argument for compilation (@sakrejda).
+* Reduce dependencies on some `rlang` functions like `sym()` (@sakrejda).
+* Change `trn()` to `if_any()` (@mattwarkentin, @sakrejda, @tjmahr).
+* Add @sakrejda and @mattwarkentin as reviewers in the `DESCRIPTION`.
+* Talk about the R package and system dependencies of `stantargets` in the README (@mattwarkentin).
+* Throw an error earlier if the Stan file does not exist (@sakrejda, @mattwarkentin)
+* Use `@format` `roxygen2` tag for data generation (@mattwarkentin).
+* Use `@family` go cross-reference functions (@mattwarkentin).
+* Elaborate on the roles and return values of specific targets generated by target factories (@mattwarkentin).
+
 # stantargets 0.0.0.9002
 
 * Return the executable file after the Stan source file in model compilation targets.
 * Replace the `log` argument with `stdout` and `stderr` (#23).
 * Switch meaning of `%||%` and `%|||%` to conform to historical precedent.
-* Add the `pedantic` argument for compilation (@sakrejda).
+
 
 # stantargets 0.0.0.9001
 
