### 2021-04-09 #430 @wlandau
@melvidoni, I believe I have implemented most of the items you identified. If you would add me as an admin of https://github.com/ropensci/stantargets, I can deactivate the current GitHub Pages setup and migrate to https://docs.ropensci.io.
