+++ NEWS.md
@@ -11,13 +11,13 @@
 * Use `@format` `roxygen2` tag for data generation (@mattwarkentin).
 * Use `@family` go cross-reference functions (@mattwarkentin).
 * Elaborate on the roles and return values of specific targets generated by target factories (@mattwarkentin).
+* Undergo rOpenSci peer review and migrate to rOpenSci.
 
 # stantargets 0.0.0.9002
 
 * Return the executable file after the Stan source file in model compilation targets.
 * Replace the `log` argument with `stdout` and `stderr` (#23).
 * Switch meaning of `%||%` and `%|||%` to conform to historical precedent.
-
 
 # stantargets 0.0.0.9001
 
