### 2021-04-12 f87d33d
Update NEWS.md
