+++ NEWS.md
@@ -11,7 +11,7 @@
 * Use `@format` `roxygen2` tag for data generation (@mattwarkentin).
 * Use `@family` go cross-reference functions (@mattwarkentin).
 * Elaborate on the roles and return values of specific targets generated by target factories (@mattwarkentin).
-* Undergo rOpenSci peer review and migrate to rOpenSci.
+* Undergo rOpenSci peer review and transition to rOpenSci.
 
 # stantargets 0.0.0.9002
 
