### 2021-04-12 #430 @wlandau
Thanks, @melvidoni. I deactivated GitHub Pages and removed the `gh-pages` branch. Does anything else need to be done for https://docs.ropensci.org/stantargets/ to appear?
