### 2021-04-14 #430 @melvidoni
> Thanks, @melvidoni. I deactivated GitHub Pages and removed the `gh-pages` branch. Does anything else need to be done for https://docs.ropensci.org/stantargets/ to appear?

Hello, apologies for the delay in coming back to you. There was a wrong GitHub token somewhere so things were stalled. I spoke with the Editors and believe the website should appear today or tomorrow.
