### 2021-04-14 #430 @jeroen
See: https://dev.ropensci.org/job/stantargets/1/console

The docs didn't appear because the fails with:

```
-- Building articles -----------------------------------------------------------
Writing 'articles/index.html'
Reading 'vignettes/mcmc.Rmd'
-- RMarkdown error -------------------------------------------------------------
Quitting from lines 120-121 (mcmc.Rmd)
Error : callr subprocess failed: CmdStan path has not been set yet. See ?set_cmdstan_path.
Visit https://books.ropensci.org/targets/debugging.html for debugging advice.
Error : Failed to render RMarkdown
```

Is this something we should install on the docs server?
