### 2021-04-14 #430 @wlandau
Ah, that explains it. I will try the `IN_PKGDOWN` workaround.

More generally, do package developers have control over custom system dependency requirements in ropensci.org builds like we do in GHA? Not sure how I feel about a system-wide CmdStan installation (it is developing rapidly) and `IN_PKGDOWN` assumes the vignettes always build first.
