### 2021-04-14 a743d13
Install cmdstanr in vignettes
