### 2021-04-14 #430 @jeroen
The image that we use for building the docs has all the same system dependencies pre-installed as on the CRAN servers. So if your package is intended to work on CRAN, it should also work on our build servers.
