### 2021-04-14 #430 @wlandau
> The image that we use for building the docs has all the same system dependencies pre-installed as on the CRAN servers. So if your package is intended to work on CRAN, it should also work on our build servers.

That's reasonable.

> You could also move this line to your package onLoad function so that it runs first.

I will consider it. So far, I have been trying to follow what `cmdstanr` does. I am not sure when `cmdstanr` will reach CRAN, and I do wonder how @jgabry and @rok-cesnovar plan to handle the dependency on CmdStan when the time comes for a release.
