### 2021-04-14 #430 @jgabry
> and I do wonder how @jgabry and @rok-cesnovar et al. plan to handle the dependency on CmdStan when the time comes for a release.

Probably similarly to how RStudio handles (or handled) it's tensorflow package. I haven't checked recently so maybe they've changed it, but it used to be that no tests involving tensorflow were run on CRAN, which is what we'd have to do with CmdStan since CRAN won't have a CmdStan installation.
