### 2021-04-14 #430 @wlandau
Yeah, and I also noticed `tensorflow` has no `vignettes/` folder. `cmdstanr` might be able to host its vignettes elsewhere, but built-in vignettes are a requirement for rOpenSci packages like `stantargets`.

Actually, I just remembered that [`stantargets` tries to suppress code in the vignettes if it detects CRAN](https://github.com/ropensci/stantargets/blob/6ff7e21f5739d41918445cb8a44be758d14f2ac1/vignettes/mcmc.Rmd#L20-L22). Hopefully that will work.

```r
if (identical(Sys.getenv("NOT_CRAN", unset = "false"), "false")) {
  knitr::opts_chunk$set(eval = FALSE)
}
```
