### 2021-04-14 #430 @jgabry
Yeah we also have some logic in the vignettes to prevent them from running on CRAN. We've used similar logic with no issues before in packages already on CRAN like rstanarm and brms where the vignettes could in theory be run on CRAN but take too long.
