### 2021-04-14 #430 @stefaniebutland
> we can maybe make a blogpost about this package?

absolutely. blog post or tech note according to Will's preference 🙂
https://blogguide.ropensci.org/
