### 2021-04-20 1f9b58f
Fix refs
https://github.com/openjournals/joss-reviews/issues/3193#issuecomment-822671328
