+++ inst/paper.bib
@@ -59,19 +59,20 @@
   year           = {2014}
 }
 
-@Manual{
+@misc{
   stantargets,
   title = {stantargets: Targets for Stan Workflows},
   author = {William Michael Landau},
-  note = {https://docs.ropensci.org/stantargets/, https://github.com/ropensci/stantargets},
+  howpublished = {\url{https://docs.ropensci.org/stantargets/}},
+  year = {2021}
 }
 
-@Manual{
+@misc{
   cmdstanr,
   title = {cmdstanr: {R} Interface to '{C}md{S}tan'},
   author = {Jonah Gabry and Rok Češnovar},
-  year = {2020},
-  note = {https://mc-stan.org/cmdstanr, https://discourse.mc-stan.org}
+  year = {2021},
+  howpublished = {\url{https://mc-stan.org/cmdstanr}}
 }
 
 @misc{
