+++ inst/paper.bib
@@ -8,7 +8,7 @@
   number = {57},
   pages = {2959},
   author = {William Michael Landau},
-  title = {The targets R package: a dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing},
+  title = {The targets {R} package: a dynamic Make-like function-oriented pipeline toolkit for reproducibility and high-performance computing},
   journal = {Journal of Open Source Software}
 }
 
@@ -61,7 +61,7 @@
 
 @misc{
   stantargets,
-  title = {stantargets: Targets for Stan Workflows},
+  title = {stantargets: targets for {S}tan workflows},
   author = {William Michael Landau},
   howpublished = {\url{https://docs.ropensci.org/stantargets/}},
   year = {2021}
