+++ README.Rmd
@@ -37,7 +37,7 @@ The `stantargets` R package is an extension to [`targets`](https://docs.ropensci
 
 ## How to get started
 
-Read the `stantargets` tutorial vignettes [here](https://docs.ropensci.org/stantargets/articles/mcmc.html) and [here](https://docs.ropensci.org/stantargets/articles/mcmc_rep.html), then use <https://docs.ropensci.org/stantargets/> as a reference while constructing your own workflows.
+Read the `stantargets` tutorial vignettes [here](https://docs.ropensci.org/stantargets/articles/mcmc.html) and [here](https://docs.ropensci.org/stantargets/articles/mcmc_rep.html), and use <https://docs.ropensci.org/stantargets/> as a reference while constructing your own workflows. Visit <https://github.com/wlandau/stantargets-example-validation> for an example project based on <https://docs.ropensci.org/stantargets/articles/mcmc_rep.html>. The example has an [RStudio Cloud workspace](https://rstudio.cloud/project/2466069) which allows you to run the project in a web browser.
 
 ## Installation
 
