+++ NEWS.md
@@ -1,4 +1,4 @@
-# stantargets 0.0.0.9003
+# stantargets 0.0.1
 
 * Skip tests if CmdStan is not installed (@sakrejda).
 * Use custom `generate_data()` function in the docs, as opposed to `tar_stan_example_data()` directly (@sakrejda).
@@ -12,6 +12,7 @@
 * Use `@family` go cross-reference functions (@mattwarkentin).
 * Elaborate on the roles and return values of specific targets generated by target factories (@mattwarkentin).
 * Undergo rOpenSci peer review and transition to rOpenSci.
+* Link to an example project.
 
 # stantargets 0.0.0.9002
 
