+++ NEWS.md
@@ -1,3 +1,7 @@
+# stantargets 0.0.1.9000
+
+
+
 # stantargets 0.0.1
 
 * Skip tests if CmdStan is not installed (@sakrejda).
