+++ NEWS.md
@@ -1,3 +1,6 @@
+# stantargets 0.0.2.9000
+
+
 # stantargets 0.0.2
 
 * Reference JOSS paper.
