### 2021-04-21 #430 @wlandau
A couple updates:

1. I got permission to release https://github.com/wlandau/stantargets-example-validation, the example `stantargets` project. (@mattwarkentin mentioned the idea in his review.) It has the example code from the [simulation study vignette](https://docs.ropensci.org/stantargets/articles/mcmc_rep.html) and an [RStudio Cloud](https://rstudio.cloud/project/2466069) workspace to let people run it in a web browser. https://github.com/wlandau/stantargets-example-validation is just like https://github.com/wlandau/targets-stan, but with a [much shorter and simpler `_targets.R` file](https://github.com/wlandau/stantargets-example-validation/blob/main/_targets.R) because of the abstractions of `stantargets`.
2. The [JOSS manuscript](https://github.com/ropensci/stantargets/blob/main/inst/paper.md) was [just approved](https://github.com/openjournals/joss-reviews/issues/3188)!
