+++ vignettes/introduction.Rmd
@@ -50,8 +50,15 @@ model {
 writeLines(lines, "x.stan")
 ```
 
+A typical workflow proceeds as follows:
 
-Next, create a `_targets.R` file to define the [`targets`](https://docs.ropensci.org/targets/) pipeline.
+1. Prepare a list of input data to Stan, including vector elements `x` and `y`.
+1. Fit the Stan model using the list of input data.
+1. Use the fitted model object to compute posterior summaries and convergence diagnostics.
+1. Use the fitted model object to extract posterior draws of parameters and store them in a tidy ata frame.
+1. Use the fitted model to compute Hamiltonian Monte Carlo (HMC) diagnostics.
+
+`stantargets` expresses this workflow using the [`tar_stan_mcmc()`](https::docs.ropensci.org/stantargets/reference/tar_stan_mcmc.html) function. To use it in a [`targets`](https://docs.ropensci.org/targets/) pipeline, invoke it from the `_targets.R` script of the project. 
 
 ```{r, echo = FALSE}
 library(targets)
@@ -94,7 +101,7 @@ list(
 )
 ```
 
-`tar_stan_mcmc(example, ...)` defines several targets:
+Above, `tar_stan_mcmc(example, ...)` only *defines* the pipeline. It does not actually run Stan, it declares the targets that will eventually run Stan. The specific targets are as follows.
 
 * `example_file_x`: Reproducibly track changes to the Stan model file.
 * `example_data`: Run the code you supplied to `data` and return a dataset compatible with Stan.
@@ -105,7 +112,7 @@ list(
 
 The suffix `_x` comes from the base name of the model file, in this case `x.stan`. If you supply multiple model files to the `stan_files` argument, all the models share the same dataset, and the suffixes distinguish among the various targets.
 
-Run `tar_manifest()` to see a tidy data frame of commands and settings for the targets.
+Run `tar_manifest()` to show specific details about the targets.
 
 ```{r}
 tar_manifest()
