+++ man/tar_stan_mcmc_rep_draws.Rd
@@ -91,7 +91,7 @@ combine all the model results
 into a single data frame downstream. Convenient, but
 duplicates data.}
 
-\item{compile}{Do compilation? The default is \code{TRUE}. If \code{FALSE}
+\item{compile}{(logical) Do compilation? The default is \code{TRUE}. If \code{FALSE}
 compilation can be done later via the \code{\link[cmdstanr:model-method-compile]{$compile()}}
 method.}
 
@@ -140,11 +140,11 @@ not modified since last compiled. The default is \code{FALSE}.}
 \item{seed}{(positive integer(s)) A seed for the (P)RNG to pass to CmdStan.
 In the case of multi-chain sampling the single \code{seed} will automatically be
 augmented by the the run (chain) ID so that each chain uses a different
-seed. The exception is the \verb{transformed data} block, which defaults to
+seed. The exception is the transformed data block, which defaults to
 using same seed for all chains so that the same data is generated for all
 chains if RNG functions are used. The only time \code{seed} should be specified
 as a vector (one element per chain) is if RNG functions are used in
-\verb{transformed data} and the goal is to generate \emph{different} data for each
+transformed data and the goal is to generate \emph{different} data for each
 chain.}
 
 \item{refresh}{(non-negative integer) The number of iterations between
@@ -152,7 +152,7 @@ printed screen updates. If \code{refresh = 0}, only error messages will be
 printed.}
 
 \item{init}{(multiple options) The initialization method to use for the
-variables declared in the \code{parameters} block of the Stan program:
+variables declared in the parameters block of the Stan program:
 \itemize{
 \item A real number \code{x>0}. This initializes \emph{all} parameters randomly between
 \verb{[-x,x]} (on the \emph{unconstrained} parameter space);
@@ -209,9 +209,9 @@ is to look for the option \code{"mc.cores"}, which can be set for an entire \R
 session by \code{options(mc.cores=value)}. If the \code{"mc.cores"} option has not
 been set then the default is \code{1}.}
 
-\item{chain_ids}{(vector) A vector of chain IDs. Must contain \code{chains} unique
-positive integers. If not set, the default chain IDs are used (integers
-starting from \code{1}).}
+\item{chain_ids}{(integer vector) A vector of chain IDs. Must contain
+\code{chains} unique positive integers. If not set, the default chain IDs are
+used (integers starting from \code{1}).}
 
 \item{threads_per_chain}{(positive integer) If the model was
 \link[cmdstanr:model-method-compile]{compiled} with threading support, the number of
@@ -256,12 +256,12 @@ statistic.}
 approximation to continuous Hamiltonian dynamics. This is further tuned
 during warmup.}
 
-\item{metric}{(character) One of \code{"diag_e"}, \code{"dense_e"}, or \code{"unit_e"},
+\item{metric}{(string) One of \code{"diag_e"}, \code{"dense_e"}, or \code{"unit_e"},
 specifying the geometry of the base manifold. See the \emph{Euclidean Metric}
 section of the CmdStan User's Guide for more details. To specify a
 precomputed (inverse) metric, see the \code{inv_metric} argument below.}
 
-\item{metric_file}{(character) A character vector containing paths to JSON or
+\item{metric_file}{(character vector) The paths to JSON or
 Rdump files (one per chain) compatible with CmdStan that contain
 precomputed inverse metrics. The \code{metric_file} argument is inherited from
 CmdStan but is confusing in that the entry in JSON or Rdump file(s) must be
@@ -316,8 +316,17 @@ fit object can be used to display all the silenced messages.}
 \item{inc_warmup}{(logical) Should warmup draws be included? Defaults to
 \code{FALSE}. Ignored except when used with \link[cmdstanr]{CmdStanMCMC} objects.}
 
[…]
