+++ R/tar_stan_gq_rep.R
@@ -47,6 +47,8 @@ tar_stan_gq_rep <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = targets::tar_option_get("memory"),
   garbage_collection = targets::tar_option_get("garbage_collection"),
@@ -175,7 +177,7 @@ tar_stan_gq_rep <- function(
     pattern = pattern_data,
     packages = packages,
     library = library,
-    format = "qs",
+    format = format,
     iteration = "list",
     error = error,
     memory = memory,
@@ -189,7 +191,7 @@ tar_stan_gq_rep <- function(
     command = command,
     pattern = pattern,
     packages = character(0),
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
