+++ R/tar_stan_mcmc_rep_diagnostics.R
@@ -110,6 +110,8 @@ tar_stan_mcmc_rep_diagnostics <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = "transient",
   garbage_collection = TRUE,
@@ -170,6 +172,8 @@ tar_stan_mcmc_rep_diagnostics <- function(
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,
+    format = format,
+    format_df = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
