+++ R/tar_stan_mcmc_rep_draws.R
@@ -112,6 +112,8 @@ tar_stan_mcmc_rep_draws <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = "transient",
   garbage_collection = TRUE,
@@ -173,6 +175,8 @@ tar_stan_mcmc_rep_draws <- function(
     tidy_eval = tidy_eval,
     packages = packages,
     library = library,
+    format = format,
+    format_df = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
