+++ R/tar_stan_gq.R
@@ -102,6 +102,8 @@ tar_stan_gq <- function(
   tidy_eval = targets::tar_option_get("tidy_eval"),
   packages = targets::tar_option_get("packages"),
   library = targets::tar_option_get("library"),
+  format = "qs",
+  format_df = "fst_tbl",
   error = targets::tar_option_get("error"),
   memory = targets::tar_option_get("memory"),
   garbage_collection = targets::tar_option_get("garbage_collection"),
@@ -199,7 +201,7 @@ tar_stan_gq <- function(
     command = command_data,
     packages = packages,
     library = library,
-    format = "qs",
+    format = format,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
@@ -210,7 +212,7 @@ tar_stan_gq <- function(
   target_output <- targets::tar_target_raw(
     name = name_gq,
     command = command_gq,
-    format = "qs",
+    format = format,
     packages = character(0),
     error = error,
     memory = memory,
@@ -226,7 +228,7 @@ tar_stan_gq <- function(
     name = name_draws,
     command = command_draws,
     packages = character(0),
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
@@ -238,7 +240,7 @@ tar_stan_gq <- function(
     name = name_summary,
     command = command_summary,
     packages = packages,
-    format = "fst_tbl",
+    format = format_df,
     error = error,
     memory = memory,
     garbage_collection = garbage_collection,
