+++ man/tar_stan_summary.Rd
@@ -3,12 +3,6 @@
 \name{tar_stan_summary}
 \alias{tar_stan_summary}
 \title{One summary of a \code{CmdStanFit} object}
-\format{
-Character of length 1, storage format of the data frame
-of posterior summaries. We recommend efficient data frame formats
-such as \code{"feather"} or \code{"aws_parquet"}. For more on storage formats,
-see the help file of \code{targets::tar_target()}.
-}
 \usage{
 tar_stan_summary(
   name,
@@ -48,12 +42,10 @@ on the \code{CmdStanFit} object.}
 \code{.args} in \code{posterior::summarize_draws()} through \verb{$summary()}
 on the \code{CmdStanFit} object.}
 
-\item{format}{Character of length 1, storage format of the non-data-frame
-targets such as the Stan data and any CmdStanFit objects.
-Please choose an all=purpose
-format such as \code{"qs"} or \code{"aws_qs"} rather than a file format like
-\code{"file"} or a data frame format like \code{"parquet"}. For more on storage
-formats, see the help file of \code{targets::tar_target()}.}
+\item{format}{Character of length 1, storage format of the data frame
+of posterior summaries. We recommend efficient data frame formats
+such as \code{"feather"} or \code{"aws_parquet"}. For more on storage formats,
+see the help file of \code{targets::tar_target()}.}
 
 \item{error}{Character of length 1, what to do if the target
 runs into an error. If \code{"stop"}, the whole pipeline stops
