+++ R/tar_stan_mle.R
@@ -358,5 +358,6 @@ tar_stan_mle_run <- function(
   # https://github.com/stan-dev/cmdstanr/blob/d27994f804c493ff3047a2a98d693fa90b83af98/R/fit.R#L16-L18 # nolint
   fit$draws() # Do not specify variables.
   try(fit$init(), silent = TRUE)
+  try(fit$profiles(), silent = TRUE)
   fit
 }
