### 2021-06-09 #430 @wlandau
On reflection, I think it might be nice to broaden the tech note (or blog post) to Bayesian data analysis pipelines in general and also include `jagstargets` when it completes its review.
