### 2021-06-17 42f2008
Suppress package installation warnings
