+++ .github/workflows/check.yaml
@@ -120,7 +120,7 @@ jobs:
       - name: Install dependencies
         run: |
           remotes::install_cran("knitr")
-          remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary"))
+          suppressWarnings(remotes::install_deps(dependencies = TRUE, type = ifelse(tolower(Sys.info()["sysname"]) == "linux", "source", "binary")))
           remotes::install_cran("rcmdcheck")
         shell: Rscript {0}
 
