### 2021-06-21 e67c190
Sync with Stan docs
