### 2021-06-21 9601658
Use newly exported targets assertion system
and other utilities
