+++ R/tar_stan_gq_rep.R
@@ -61,8 +61,8 @@ tar_stan_gq_rep <- function(
 ) {
   envir <- tar_option_get("envir")
   compile <- match.arg(compile)
-  assert_chr(stan_files)
-  assert_unique(stan_files)
+  targets::tar_assert_chr(stan_files)
+  targets::tar_assert_unique(stan_files)
   lapply(stan_files, assert_stan_file)
   name_stan <- produce_stan_names(stan_files)
   name_file <- paste0(name, "_file")
@@ -75,7 +75,7 @@ tar_stan_gq_rep <- function(
   sym_batch <- as.symbol(name_batch)
   sym_data <- as.symbol(name_data)
   command_batch <- substitute(seq_len(x), env = list(x = batches))
-  command_rep <- tar_tidy_eval(
+  command_rep <- targets::tar_tidy_eval(
     data,
     envir = envir,
     tidy_eval = tidy_eval
