+++ R/tar_stan_mcmc_rep_draws.R
@@ -125,7 +125,7 @@ tar_stan_mcmc_rep_draws <- function(
   cue = targets::tar_option_get("cue")
 ) {
   tar_stan_mcmc_rep(
-    name = deparse_language(substitute(name)),
+    name = targets::tar_deparse_language(substitute(name)),
     stan_files = stan_files,
     data = substitute(data),
     output_type = "draws",
