### 2021-06-21 4a1948f
Relax assertion
