+++ R/utils_assert.R
@@ -1,11 +1,4 @@
 assert_stan_file <- function(stan_file) {
-  targets::tar_assert_scalar(
-    stan_file,
-    paste(
-      "stan_file must have length 1.",
-      "Some stantargets functions do allow multiple Stan files though."
-    )
-  )
   targets::tar_assert_chr(stan_file)
   targets::tar_assert_path(stan_file)
   if (dir.exists(stan_file)) {
