+++ R/utils_stan.R
@@ -4,8 +4,14 @@ produce_stan_names <- function(stan_files) {
     fs::path_ext_remove(basename(stan_files)),
     names(stan_files)
   )
-  targets::tar_assert_unique(out, "target suffixes from stan_files are not unique.")
-  targets::tar_assert_nzchar(out, "target suffixes from stan_files must be nonempty.")
+  targets::tar_assert_unique(
+    out,
+    "target suffixes from stan_files are not unique."
+  )
+  targets::tar_assert_nzchar(
+    out,
+    "target suffixes from stan_files must be nonempty."
+  )
   make.names(out)
 }
 
