+++ R/tar_stan_package.R
@@ -24,7 +24,8 @@
 #' @importFrom rlang check_installed expr quo_squash
 #' @importFrom stats rnorm runif
 #' @importFrom targets tar_assert_chr tar_assert_nonempty
-#'   tar_assert_not_in tar_assert_nzchar tar_assert_path
+#'   tar_assert_not_dir tar_assert_not_in
+#'   tar_assert_nzchar tar_assert_path
 #'   tar_assert_scalar tar_assert_unique
 #'   tar_cue tar_deparse_safe tar_dir tar_load tar_option_get tar_path
 #'   tar_read tar_script tar_target tar_target_raw tar_test tar_tidy_eval
