+++ R/utils_assert.R
@@ -1,7 +1,5 @@
 assert_stan_file <- function(stan_file) {
   targets::tar_assert_chr(stan_file)
   targets::tar_assert_path(stan_file)
-  if (dir.exists(stan_file)) {
-    targets::tar_throw_validate("Stan model file must not be a directory.")
-  }
+  targets::tar_assert_not_dir(stan_file)
 }
